Module Catala.Context

type uid = Uid.t

type scope_uid = Uid.t

type var_uid = Uid.t

type sub_scope_uid = Uid.t

module UidMap = Uid.UidMap

module UidSet = Uid.UidSet

type ident = string

module IdentMap : sig ... end

type typ = Lambda.typ

type sort =

| IdScope

| IdScopeVar of uid option

| IdSubScope of uid

| IdSubScopeVar of var_uid * sub_scope_uid

| IdBinder

type uid_data = {

uid_typ : typ;

uid_sort : sort;

}

type scope_context = {

var_id_to_uid : uid IdentMap.t;

uid_set : UidSet.t;

}

type context = {

scope_id_to_uid : uid IdentMap.t;

scopes : scope_context UidMap.t;

data : uid_data UidMap.t;

}

val subscope_ident : string -> string -> string

val raise_unsupported_feature : string -> Pos.t -> 'a

val raise_undefined_identifier : string -> Pos.t -> 'a

val raise_unknown_identifier : string -> Pos.t -> 'a

val get_uid_typ : context -> uid -> typ

Get the type associated to an uid

val get_uid_sort : context -> uid -> sort

Get the sort associated to an uid

val process_subscope_decl : uid -> context -> Ast.scope_decl_context_scope -> context

Process a subscope declaration

val process_base_typ : Ast.base_typ Pos.marked -> Lambda.typ

val process_type : Ast.typ Pos.marked -> Lambda.typ

val process_data_decl : uid -> context -> Ast.scope_decl_context_data -> context

Process data declaration

val process_item_decl : uid -> context -> Ast.scope_decl_context_item -> context

Process an item declaration

val process_scope_decl : context -> Ast.scope_decl -> context

Process a scope declaration

val process_code_item : context -> Ast.code_item -> context

Process a code item : for now it only handles scope decls

val process_code_block : context -> Ast.code_block -> context

Process a code block

val process_program_item : context -> Ast.program_item -> context

Process a program item

val form_context : Ast.program -> context

Derive the context from metadata

val get_var_uid : uid -> context -> ident Pos.marked -> uid

Get the variable uid inside the scope given in argument

val get_subscope_uid : uid -> context -> ident Pos.marked -> uid * uid

Get the subscope uid inside the scope given in argument

val belongs_to : context -> var_uid -> scope_uid -> bool

Checks if the var_uid belongs to the scope scope_uid

val add_binding : context -> Uid.t -> Uid.t -> ident Pos.marked option -> context * Uid.t option

Adds a binding to the context