Module Catala.Firstpass

val subscope_ident : string -> string -> string

val expr_to_lambda : ?⁠subdef:Uid.t -> Uid.t -> Context.context -> Ast.expression Pos.marked -> Lambda.term

The optional argument subdef allows to choose between differents uids in case the expression is a redefinition of a subvariable

val typing : Context.context -> Lambda.term -> Lambda.term * Lambda.typ

Checks that a term is well typed and annotate it

val merge_conditions : Lambda.term option -> Lambda.term option -> Pos.t -> Lambda.term

val process_def : Lambda.term option -> Uid.t -> Context.context -> Scope.program -> Ast.definition -> Scope.program

val process_rule : Lambda.term option -> Uid.t -> Context.context -> Scope.program -> Ast.rule -> Scope.program

Process a rule from the surface language

val process_scope_use_item : Lambda.term option -> Uid.t -> Context.context -> Scope.program -> Ast.scope_use_item Pos.marked -> Scope.program

val process_scope_use : Context.context -> Scope.program -> Ast.scope_use -> Scope.program

val translate_program_to_scope : Context.context -> Ast.program -> Scope.program

Scopes processing