Module Catala__Firstpass

val subscope_ident : string -> string -> string

val expr_to_lambda : ?⁠subdef:Catala.Uid.t -> Catala.Uid.t -> Catala.Context.context -> Catala.Ast.expression Catala.Pos.marked -> Catala.Lambda.term

The optional argument subdef allows to choose between differents uids in case the expression is a redefinition of a subvariable

val typing : Catala.Context.context -> Catala.Lambda.term -> Catala.Lambda.term * Catala.Lambda.typ

Checks that a term is well typed and annotate it

val merge_conditions : Catala.Lambda.term option -> Catala.Lambda.term option -> Catala.Pos.t -> Catala.Lambda.term

val process_def : Catala.Lambda.term option -> Catala.Uid.t -> Catala.Context.context -> Catala.Scope.program -> Catala.Ast.definition -> Catala.Scope.program

val process_rule : Catala.Lambda.term option -> Catala.Uid.t -> Catala.Context.context -> Catala.Scope.program -> Catala.Ast.rule -> Catala.Scope.program

Process a rule from the surface language

val process_scope_use_item : Catala.Lambda.term option -> Catala.Uid.t -> Catala.Context.context -> Catala.Scope.program -> Catala.Ast.scope_use_item Catala.Pos.marked -> Catala.Scope.program

val process_scope_use : Catala.Context.context -> Catala.Scope.program -> Catala.Ast.scope_use -> Catala.Scope.program

val translate_program_to_scope : Catala.Context.context -> Catala.Ast.program -> Catala.Scope.program

Scopes processing