Module Catala__Html

This modules weaves the source code and the legislative text together into a document that law professionals can understand.

module A = Catala.Ast

module P = Stdlib.Printf

module R = Re.Pcre

module C = Catala.Cli

val pre_html : string -> string

val raise_failed_pygments : string -> int -> 'a

val wrap_html : string -> string list -> string option -> Catala.Cli.backend_lang -> string

val pygmentize_code : string Catala.Pos.marked -> C.backend_lang -> string option -> string

type program_state =

| InsideArticle

| OutsideArticle

val program_item_to_html : A.program_item -> string option -> C.backend_lang -> program_state -> string * program_state

val ast_to_html : A.program -> string option -> C.backend_lang -> string