Module Catala__Interpreter

module G = Graph.Pack.Digraph

type uid = int

type scope_uid = int

module UidMap = Catala.Uid.UidMap

module UidSet = Catala.Uid.UidSet

type exec_context = Catala.Lambda.untyped_term UidMap.t

val empty_exec_ctxt : 'a UidMap.t

val substitute : uid -> Catala.Lambda.untyped_term -> Catala.Lambda.term -> Catala.Lambda.term

val eval_term : exec_context -> Catala.Lambda.term -> Catala.Lambda.term

val eval_default_term : exec_context -> Catala.Lambda.default_term -> (Catala.Lambda.term, Catala.Pos.t list * Catala.Pos.t list) Stdlib.result

val build_scope_schedule : Catala.Context.context -> Catala.Scope.scope -> G.t

Returns the scheduling of the scope variables, if y is a subscope and x a variable of y, then we have two different variable y.x(internal) and y.x(result) and the ordering y.x(internal) -> y -> y.x(result)

val merge_var_redefs : Catala.Scope.scope -> Catala.Scope.definition UidMap.t -> Catala.Scope.scope

val raise_default_conflict : string Catala.Pos.marked -> Catala.Pos.t list -> Catala.Pos.t list -> 'a

val execute_scope : ?⁠exec_context:Catala.Lambda.untyped_term UidMap.t -> Catala.Context.context -> Catala.Scope.program -> Catala.Scope.scope -> exec_context