Module Catala_runtime.Print

This module is for setting options and internals, use Value.format to print values

type lang = [

| `En

| `Fr

| `Pl

]

val set_lang : [ `En | `Fr | `Pl ] -> unit

Sets the language to be used for the output of values

val get_lang : unit -> [ `En | `Fr | `Pl ]

val set_precision : int -> unit

Sets the maximum number of decimal numbers to print

val get_precision : unit -> int