Module Catala_runtime.Value

type _ external_tag = ..

module type External = sig ... end

type _ ty =

| Unit : unit ty

| Bool : bool ty

| Integer : integer ty

| Money : money ty

| Decimal : decimal ty

| Date : date ty

| Duration : duration ty

| Position : code_location ty

| Array : ('a -> t) -> 'a array ty

| Tuple : ('a -> t list) -> 'a ty

| Struct : {

name : string;

fields : 'a -> (string * t) list;

} -> 'a ty

| Enum : {

name : string;

constr : 'a -> int * string * t option;

} -> 'a ty

| External : (module External with type t = 'a) -> 'a ty

| Function : 'a ty

| Polymorphic : 'a ty

'a ty provides runtime information about the structure of values of OCaml type 'a

and t =

| V : 'a ty * 'a -> t

Runtime.Value.t is an embedded runtime value that comes with type information, allowing for introspection

val embed : 'a ty -> 'a -> t

val equal : code_location -> t -> t -> bool

val compare : code_location -> t -> t -> int

val format : Stdlib.Format.formatter -> t -> unit

val from_json : 'a ty -> code_location -> string -> 'a