Module Catala_utils

module Cli : sig ... end

module File : sig ... end

module Global : sig ... end

This module contains definitions of global flags and types used throughout. They should be defined from the command-line and never modified afterwards.

module Graphs : sig ... end

module Hash : sig ... end

Hashes for the identification of modules.

module Map : sig ... end

Small wrapper on top of the stdlib Map module to add some useful functions

module Mark : sig ... end

AST node annotations (used for position, type, etc.)

module Message : sig ... end

Interface for emitting compiler messages.

module Pos : sig ... end

Source code position

module String : sig ... end

module Suggestions : sig ... end

module Uid : sig ... end

Global identifiers factories using a generative functor

module Version : sig ... end

The implementation of this module is generated by the build system (through get_version.ml), and should only contain a static string