Module Catala_utils.Cli

type backend_lang =

| En

| Fr

| Pl

type when_enum =

| Auto

| Always

| Never

The usual auto/always/never option argument

type message_format_enum =

| Human

| GNU	(*
Format of error and warning messages output by the compiler.
*)

type input_file =

| FileName of string

| Contents of string

val languages : (string * backend_lang) list

val language_code : backend_lang -> string

Returns the lowercase two-letter language code

Configuration globals

type options = private {

mutable input_file : input_file;

mutable language : backend_lang option;

mutable debug : bool;

mutable color : when_enum;

mutable message_format : message_format_enum;

mutable trace : bool;

mutable plugins_dirs : string list;

mutable disable_warnings : bool;

mutable max_prec_digits : int;

}

Global options, common to all subcommands (note: the fields are internally mutable only for purposes of the globals toplevel value defined below)

val globals : options

A global definition to the global options is provided for convenience, e.g. choosing the proper output in formatting functions. Prefer the use of the options returned by the command-line parsing whenever possible.

val enforce_globals : ?input_file:input_file -> ?language:backend_lang option -> ?debug:bool -> ?color:when_enum -> ?message_format:message_format_enum -> ?trace:bool -> ?plugins_dirs:string list -> ?disable_warnings:bool -> ?max_prec_digits:int -> unit -> options

Sets up the global options (side-effect); for specific use-cases only, this should never be called from the compiler or when going through normal command-line parsing. Proper uses include setting up the compiler library when using it directly through a specific front-end.

CLI flags and options

module Flags : sig ... end

Command-line application

val version : string

val info : Cmdliner.Cmd.info

val s_plugins : string

Manpage section name for the installed plugins

exception Exit_with of int

Exit with a specific exit code (but less brutally than Sys.exit which would bypass all finalisers)