Module Flags.Global

val flags : Global.options Cmdliner.Term.t

Global flags available to all commands. Note that parsing this term also performs some side-effects into GlobalRefs and sets up signal/error processing. Sets input_src to Stdin "-stdin-", use options for the full parser

val options : Global.options Cmdliner.Term.t

flags plus an additional positional argument for the input file