Module Catala_utils.Hash

Hashes for the identification of modules.

In contrast with OCaml's basic `Hashtbl.hash`, they process the full depth of terms. Any meaningful interface change in a module should only be in hash collision with a 1/2^30 probability.

type t = private int

Native Hasthbl.hash hashes, value is truncated to 30 bits whatever the architecture (positive 31-bit integers)

type full

A "full" hash includes the Catala version and compilation flags, alongside the module interface

val raw : 'a -> t

Hashtbl.hash. Do not use on deep types (it has a bounded depth), use specific hashing functions.

module Op : sig ... end

val option : ('a -> t) -> 'a option -> t

val list : ('a -> t) -> 'a list -> t

val map : (('k -> 'v -> t -> t) -> 'map -> t -> t) -> ('k -> t) -> ('v -> t) -> 'map -> t

map fold_f key_hash_f value_hash_f map computes the hash of a map. The first argument is expected to be a Foo.Map.fold function. The result is independent of the ordering of the map.

val finalise : t -> monomorphize_types:bool -> full

Turns a raw interface hash into a full hash, ready for printing

val to_string : full -> string

val format : Stdlib.Format.formatter -> full -> unit

val of_string : string -> full

Raises Failure

val external_placeholder : string

It's inconvenient to need hash updates on external modules. This string is uses as a hash instead for those cases.

NOTE: This is a temporary solution A future approach could be to have Catala generate a module loader (with the proper hash), relieving the user implementation from having to do the registration.