Module Hash.Op

val (!) : 'a -> t

Shortcut to raw. Same warning: use with an explicit type annotation !(foo: string) to ensure it's not called on types that are recursive or include annotations.

Hint: we use !`Foo as a fancy way to generate constants for discriminating constructions

val (%) : t -> t -> t

Safe combination of two hashes (non commutative or associative, etc.)