Module Uid.Gen

Shortcut for creating a kind of uids over marked strings

Parameters

module _ : Style

Signature

type t

type info = MarkedString.info

val fresh : ?from:t -> info -> t

if from is set, the new t will set its original info from it

val get_info : t -> info

val map_info : (info -> info) -> t -> t

val compare : t -> t -> int

val equal : t -> t -> bool

val format : Stdlib.Format.formatter -> t -> unit

val to_string : t -> string

val id : t -> int

Returns the unique ID of the identifier

val hash : t -> Hash.t

While id returns a unique ID valable for a given Uid instance within a given run of catala, this is a raw hash of the identifier string. Therefore, it may collide within a given program, but remains meaninful across separate compilations.

val original_info : t -> info

Retrieve the original id info, unaffected by renamings, etc.

val original_string : t -> string

Retrieve the original id string, unaffected by renamings, etc.

val format_original : Stdlib.Format.formatter -> t -> unit

module Set : Stdlib.Set.S with type elt = t

module Map : Map.S with type key = t