Module Uid.Make

This is the generative functor that ensures that two modules resulting from two different calls to Make will be viewed as different types t by the OCaml typechecker. Prevents mixing up different sorts of identifiers.

Parameters

module X : Info

Signature

type t

type info = X.info

val fresh : info -> t

val get_info : t -> info

val compare : t -> t -> int

val equal : t -> t -> bool

val format_t : Stdlib.Format.formatter -> t -> unit

val hash : t -> int

module Set : Stdlib.Set.S with type elt = t

module SetLabels : Stdlib.MoreLabels.Set.S with type elt = t and type t = Set.t

module Map : Stdlib.Map.S with type key = t

module MapLabels : Stdlib.MoreLabels.Map.S with type key = t and type 'a t = 'a Map.t