Module Clerk_backend

val default_flags : code_coverage:bool -> trace:bool -> inplace:bool -> config:Clerk_cli.config -> Clerk_utils.Var.bindings

module type S = sig ... end

type t = (module S)

Functions useful for the backend rules definitions. ~name is for the backend name here, for convenience

val static_base_rules : Ninja_utils.def list

val extern_src : filename:string -> name:string -> ext:string -> missing:string list -> string * string list

The missing argument is simply an accumulator, returned as snd with any missing files added

We use 3 types of pseudo-targets for compiled objets. The <modname> below is the normalised module name (using String.to_id)

@catala-obj/<modname> (catala_obj_target) is what is needed for Catala evaluation (an OCaml cmxs)

@<backend>/interface/<modname> (module_target) is the possibly compiled interface that dependent modules will need to be compiled (e,g, .h, .cmi)

@<backend>/obj/<modname> or @<backend>/obj/<filename> (obj_target) is the compiled object for linking, including all its dependencies. This allows transitive compilation of required objects before linking

val catala_obj_target : string -> Ninja_utils.Expr.elt

From a module name, gives the pseudo-target that builds the Catala object required for interpretation (i.e. actually the OCaml dynlink object, .cmxs or .cmo)

val target : ?name:string -> string -> Ninja_utils.Expr.elt

val interface_dep : name:string -> string -> Ninja_utils.Expr.elt

backend name, module name -> target name

val src_dep : name:string -> string -> Ninja_utils.Expr.elt

backend name, module name -> target name

val obj_dep : name:string -> Clerk_utils.Scan.item -> Ninja_utils.Expr.elt

backend name, source item -> target name

val catala_obj_dep : Clerk_utils.Scan.item -> Ninja_utils.Expr.elt

Pseudo-target that builds the item runtime dependencies for interpretation (cmxs for modules, or cmxs of all their dependencies for other items)

module Make_backend (A : sig ... end) : sig ... end

val register : t -> unit

val get : Clerk_config.backend -> t

val all : unit -> t list

val name : t -> string

val id : t -> Clerk_config.backend

module OCaml : sig ... end

module C : sig ... end

module Java : sig ... end

module Python : sig ... end

module Backend_paths : sig ... end

/

module Flags : sig ... end