Module Clerk_backend.C

val linking_command : build_dir:string -> var_bindings:Clerk_utils.Var.bindings -> ('a -> Clerk_utils.Scan.item list) -> 'a -> string -> string list

linking_command is a that returns a list of parameter that represents a command to link all the files that were generated by the Ninja file. This command is needed by the Clerk_driver modules and not the clerk_rules modules that's why it doesn't appear in the Backend common interface.

val run_artifact : test:bool -> ?scope:string -> ?quiet:bool -> string -> int * string list

run_artifact runs the command generated by linking_command. This command is needed by the Clerk_driver modules and not the clerk_rules modules that's why it doesn't appear in the Backend common interface.

val name : string

val src_extensions : string list

source files for this backend, as generated by Catala. File extensions, dot excluded

val module_extensions : string list

possibly compiled interface files: this is the requirement for building dependent modules, e.g. .h for C, .cmi for OCaml. File extensions, dot excluded

val obj_extension : string

the main target extension (dot excluded) of objects for this backend. building this should make linking the program possible

val all_obj_extensions : string list

extensions for all compiled objects by this backend, dot excluded

val var_defs : variables:(string * string list) list -> autotest:bool -> use_default_flags:bool -> test_flags:string list -> include_dirs:string list -> Clerk_utils.Var.bindings

Returns a list of variables assigned for the Ninja file generated. Those variables can be the compiler exec of your backends, flags for a specific command, include ...

~variables is the list of already defined variables, if a variables is already defined it will takes this value instead of the one given

~autotest if the flag is activated, add the autotest flag to the catala command, each call to the catala command will come with autotest

~use_default_flags flag to tell if the catala command should only use the default flag or could had some more. The default flag being -O for optimize

~test_flags is a list of tests that will be passed to the catala interpreter tests if called

~include_dirs the list of the different direcory to include each time, usually this comes from the config in the clerk.toml file.

val rules : Clerk_backend__.Sig.Nj.def list

List of base Ninja rules needed by the backend

val build_runtime : config:Clerk_cli.config -> stdbase:string -> Clerk_backend__.Sig.Nj.def list

Returns a set of Ninja rules, in charge of building "@<name>/runtime/src" and "@<name>/runtime/obj"

val catala : ?vars:Clerk_utils.Var.bindings -> is_stdlib:bool -> inputs:Clerk_backend__.Sig.Nj.Expr.t -> implicit_in:Clerk_backend__.Sig.Nj.Expr.t -> has_scope_tests:bool -> Clerk_backend__.Sig.Nj.def Stdlib.Seq.t

This functions returns the Ninja build statements that call the catala compiler to create the target file in the backend language.

val build_object : include_dirs:string list -> same_dir_modules:(string * string) list -> Clerk_utils.Scan.item -> Clerk_backend__.Sig.Nj.def Stdlib.Seq.t

Returns a set of rules that generate the obj_extensions compiled form of a the target from the src_extensions generated by the catala rule, using the backend compiler. Returns an empty sequence on non-compiled languages.

val runtime_dir : Catala_utils.File.t Stdlib.Lazy.t

Origin directory for the runtime source on the system (based on Poll.runtime_dir)

val stdlib_subdir : Catala_utils.File.t

Subdirectory of the backend directory in "_build/libcatala" that holds the standard library files. Usually just the empty string.

val write_target_def_file : config:Clerk_cli.config -> dir:Catala_utils.File.t -> Clerk_config.target -> unit

Called on the target directory under _targets/<backend>/ after the files have been copied in

val install_runtime : config:Clerk_cli.config -> unit

Sets up the runtime and stdlib to the appropriate target dir

type Clerk_config.backend +=

| T

val config_backend : Clerk_config.backend

val current_target : Clerk_utils.Scan.item -> string -> Clerk_backend__.Sig.Nj.Expr.elt

Returns the current target (in the Clerk file context) with the given extension

val runtime_targets : only_source:bool -> string list

val interface_dep : string -> Clerk_backend__.Sig.Nj.Expr.elt

Returns the pseudo-target for generated interface of the given module, possibly compiled (ie. .cmi in the case of OCaml)

val src_dep : string -> Clerk_backend__.Sig.Nj.Expr.elt

Returns the pseudo-target for generated source of the given module

val obj_dep : Clerk_utils.Scan.item -> Clerk_backend__.Sig.Nj.Expr.elt

Returns the pseudo-target for the compiled object of the given item

val external_copy : Clerk_utils.Scan.item -> Clerk_backend__.Sig.Nj.def Stdlib.Seq.t

external_copy item is the set of rules to handle a catala module marked as external in item, most of the time the external_copy function verifies if the implementation of the external module exists and then copies it in the build directory along others modules