Module Backend.Flags

val default : variables:(string * string list) list -> autotest:bool -> use_default_flags:bool -> test_flags:string list -> include_dirs:string list -> (Clerk_utils.Var.t * string list) list

defaut ~variables ~autotest ~use_default_flags ~test_flags ~include_dirs returns a list of variables assigned for the Ninja file generated. Those variables can be the compiler exec of your backends, flags for a specific command, include ...

~variables is the list of already defined variables, if a variables is already defined it will takes this value instead of the one given

~autotest if the flag is activated, add the autotest flag to the catala command, each call to the catala command will come with autotest

~use_default_flags flag to tell if the catala command should only use the default flag or could had some more. The default flag being -O for optimize

~test_flags is a list of tests that will be passed to the catala interpreter tests if called

~include_dirs the list of the different direcory to include each time, usually this comes from the config in the clerk.toml file.