Module Clerk_backends.Python

val linking_command : build_dir:string -> ('a -> Clerk_utils.Scan.item list) -> 'a -> string -> string list

linking_command is a that returns a list of parameter that represents a command to link all the files that were generated by the Ninja file. This command is needed by the Clerk_driver modules and not the clerk_rules modules that's why it doesn't appear in the Backend common interface.

val run_artifact : Clerk_lib.Clerk_cli.config -> test:bool -> ?scope:string -> var_bindings:(Clerk_utils.Var.t * string list) list -> string -> int

run_artifact runs the command generated by linking_command. This command is needed by the Clerk_driver modules and not the clerk_rules modules that's why it doesn't appear in the Backend common interface.

module Backend : Backend.S