Module Clerk_utils

module Diff : sig ... end

A simple diffing algorithm

module Ninja : sig ... end

module Poll : sig ... end

This module scans the system for the location of various required tools and libraries, to be used as default values. (think ./configure)

module Scan : sig ... end

This module is responsible for scanning Catala files, extracting dependency and test information. It is based on the lightweight "line-parser" (Surface.Parser_driver.line)

module Var : sig ... end