Module type Diff.S

type item

The type of the item that will be compared.

type diff =

| Deleted of item array

| Added of item array

| Equal of item array

Represents the change or lack of change in a line or character between the old and new version.

type t = diff list

List of diffs which is the return value of the main function.

val get_diff : item array -> item array -> t

Returns a list of diffs between two arrays

val recover_input : t -> item array * item array

Recover the original input passed to get_diff.