Module Clerk_utils.Module_graph

type module_info = {

name : string Catala_utils.Mark.pos;

item : Scan.item;

targets : Catala_utils.String.Set.t;

}

type info = {

var_bindings : Var.bindings;

modules_map : module_info Catala_utils.String.Map.t;

targets_map : Clerk_config.target Catala_utils.String.Map.t;

linking_deps : Scan.item -> string list;
(*
item -> modules, topologically ordered
*)

inclusion_map : Scan.item Catala_utils.String.Map.t;
(*
Map from source file names to a module that includes them, for all source files that actually get included
*)

}

Info passed to the callback that shall conclude the Ninja file, once the whole file tree has been crawled. The modules and targets map differ from the raw configuration information:

the Stdlib target is added

target contents are all modules to actually include in a given target

target dependencies are flattened

val empty_info : info

val stdlib_target_name : string

val organise_modules : config:Clerk_cli.config -> var_bindings:Var.bindings -> Scan.item Stdlib.Seq.t -> info

Takes a (memoised) sequence of items from the stdlib and project tree; computes the dependency graphs of targets and modules, and returns a callback_info

val module_backends : info -> string -> Clerk_config.backend list

Returns the list of backends supported by a given module by analysing the clerk targets it belongs to, or, if none, the clerk targets of its dependencies