Module Clerk_utils.Poll

This module scans the system for the location of various required tools and libraries, to be used as default values. (think ./configure)

This module is sensitive to the CWD at first use. Therefore it's expected that chdir has been run beforehand to the project root.

val clerk_exe : Catala_utils.File.t Stdlib.Lazy.t

val catala_exe : Catala_utils.File.t Stdlib.Lazy.t

val ocaml_libdir : Catala_utils.File.t Stdlib.Lazy.t

Locates the main lib directory containing the OCaml libs

val runtime_dir : Catala_utils.File.t Stdlib.Lazy.t

Locates the directory containing the runtimes to link to (in one subdirectory per backend)

val stdlib_dir : Catala_utils.File.t Stdlib.Lazy.t

val catala_source_tree_root : Catala_utils.File.t option Stdlib.Lazy.t

Set if, and only if the current build happens from within the Catala compiler source tree itself. This requires some specific options and includes for correct bootstrapping, and running the tests without requiring an installation first.