Module Clerk_utils.Scan

This module is responsible for scanning Catala files, extracting dependency and test information. It is based on the lightweight "line-parser" (Surface.Parser_driver.line)

type item = {

file_name : Catala_utils.File.t;

module_def : string Catala_utils.Mark.pos option;

is_stdlib : bool;

extrnal : bool;

used_modules : string Catala_utils.Mark.pos list;

included_files : Catala_utils.File.t Catala_utils.Mark.pos list;

has_inline_tests : bool;

has_scope_tests : bool Stdlib.Lazy.t;

}

Contains all the data extracted from a single Catala file. Lists are in reverse file order.

val libcatala : Catala_utils.File.t

Subdirectory of the build directory holding the runtime and standard library and its built artifacts

val get_lang : Catala_utils.File.t -> Catala_utils.Global.backend_lang option

Guesses Catala dialect from file-name and global options

val catala_file : Catala_utils.File.t -> Catala_utils.Global.backend_lang -> item

Scans a single Catala file into an item

val tree : Catala_utils.File.t -> (Catala_utils.File.t * Catala_utils.File.t list * item list) Stdlib.Seq.t

Recursively scans a directory, and returns the corresponding subdirectories and items in sequence, by directory.

val test_command_args : string -> string option

Parses a test command-line (in the form "$ catala <args>") and returns the arguments as a string, or None if there is no match

val find_test_scope : lang:Catala_utils.Global.backend_lang -> Catala_utils.File.t -> bool

Checks if the given file contains #test scope annotations, recursively through file includes. The file extension takes precendence over the ~lang argument.

val target_basename : item -> Catala_utils.File.t

Returns the expected basename (without directory or extension) for artifacts based on this file: the module name is used if defined, otherwise the original file is used. In both case, it is normalised using String.to_id.

val target_file_name : item -> Catala_utils.File.t

Like target_basename, but returns a relative filename to the build directory, without extension