Module Var.Op

val (!) : string t -> string

Alias to Var.ref. Run-time reference to the given variable !var = "${xvarname}", for inclusion in strings

val (!!) : 'a t -> Ninja_utils.Expr.elt

Run-time reference to the given variable !var = "${xvarname}" as an expression element that will be appropriately quoted. This handles specific quoting rules of the input and output variables by Ninja