Module Concolic

module Ast = Scalc.Ast

module Runtime = Runtime_ocaml.Runtime

module D = Dcalc.Ast

module L = Lcalc.Ast

val format_lit : Stdlib.Format.formatter -> Shared_ast.lit Catala_utils.Marked.pos -> unit

val format_log_entry : Stdlib.Format.formatter -> Shared_ast.log_entry -> unit

val format_op : Stdlib.Format.formatter -> Scalc.Ast.operator Catala_utils.Marked.pos -> unit

val format_uid_list : Stdlib.Format.formatter -> Catala_utils.Uid.MarkedString.info list -> unit

val format_string_list : Stdlib.Format.formatter -> string list -> unit

val avoid_keywords : string -> string

val format_struct_name : Stdlib.Format.formatter -> Shared_ast.StructName.t -> unit

val format_struct_field_name : Stdlib.Format.formatter -> Shared_ast.StructField.t -> unit

val format_enum_name : Stdlib.Format.formatter -> Shared_ast.EnumName.t -> unit

val format_enum_cons_name : Stdlib.Format.formatter -> Shared_ast.EnumConstructor.t -> unit

val typ_needs_parens : Shared_ast.typ -> bool

val format_typ : Stdlib.Format.formatter -> Shared_ast.typ -> unit

val format_name_cleaned : Stdlib.Format.formatter -> string -> unit

module StringMap : sig ... end

module IntMap : sig ... end

val string_counter_map : int IntMap.t StringMap.t Stdlib.ref

For each `VarName.t` defined by its string and then by its hash, we keep track of which local integer id we've given it. This is used to keep variable naming with low indices rather than one global counter for all variables. TODO: should be removed when https://github.com/CatalaLang/catala/issues/240 is fixed.

val format_var : Stdlib.Format.formatter -> Scalc.Ast.VarName.t -> unit

val format_func_name : Stdlib.Format.formatter -> Scalc.Ast.FuncName.t -> unit

val format_var_name : Stdlib.Format.formatter -> Scalc.Ast.VarName.t -> unit

val needs_parens : Scalc.Ast.expr -> bool

val format_exception : Stdlib.Format.formatter -> Shared_ast.except Catala_utils.Marked.pos -> unit

val format_expression : Shared_ast.decl_ctx -> Stdlib.Format.formatter -> Scalc.Ast.expr -> unit

val format_statement : Shared_ast.decl_ctx -> Stdlib.Format.formatter -> Scalc.Ast.stmt Catala_utils.Marked.pos -> unit

val format_block : Shared_ast.decl_ctx -> Stdlib.Format.formatter -> Scalc.Ast.block -> unit

val format_ctx : Scopelang.Dependency.TVertex.t list -> Stdlib.Format.formatter -> Shared_ast.decl_ctx -> unit

val format_program : Stdlib.Format.formatter -> Ast.program -> Scopelang.Dependency.TVertex.t list -> unit

val name : string

val extension : string

val apply : source_file:'a -> output_file:string option -> scope:'b -> Ast.program -> Scopelang.Dependency.TVertex.t list -> unit