Module Dcalc.Ast

module ScopeName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module StructName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module StructFieldName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module StructMap : Stdlib.Map.S with type key = StructName.t

module EnumName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module EnumConstructor : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module EnumMap : Stdlib.Map.S with type key = EnumName.t

Abstract syntax tree for the default calculus

Abstract syntax tree

type typ_lit =

| TBool

| TUnit

| TInt

| TRat

| TMoney

| TDate

| TDuration

type typ =

| TLit of typ_lit

| TTuple of typ Utils.Pos.marked list * StructName.t option

| TEnum of typ Utils.Pos.marked list * EnumName.t

| TArrow of typ Utils.Pos.marked * typ Utils.Pos.marked

| TArray of typ Utils.Pos.marked

| TAny

type date = Runtime.date

type duration = Runtime.duration

type lit =

| LBool of bool

| LEmptyError

| LInt of Runtime.integer

| LRat of Runtime.decimal

| LMoney of Runtime.money

| LUnit

| LDate of date

| LDuration of duration

type op_kind =

| KInt

| KRat

| KMoney

| KDate

| KDuration	(*
All ops don't have a KDate and KDuration.
*)

type ternop =

| Fold

type binop =

| And

| Or

| Xor

| Add of op_kind

| Sub of op_kind

| Mult of op_kind

| Div of op_kind

| Lt of op_kind

| Lte of op_kind

| Gt of op_kind

| Gte of op_kind

| Eq

| Neq

| Map

| Concat

| Filter

type log_entry =

| VarDef of typ	(*
During code generation, we need to know the type of the variable being logged for embedding
*)

| BeginCall

| EndCall

| PosRecordIfTrueBool

type unop =

| Not

| Minus of op_kind

| Log of log_entry * Utils.Uid.MarkedString.info list

| Length

| IntToRat

| GetDay

| GetMonth

| GetYear

type operator =

| Ternop of ternop

| Binop of binop

| Unop of unop

type expr =

| EVar of expr Bindlib.var Utils.Pos.marked

| ETuple of expr Utils.Pos.marked list * StructName.t option	(*
The MarkedString.info is the former struct field name
*)

| ETupleAccess of expr Utils.Pos.marked * int * StructName.t option * typ Utils.Pos.marked list	(*
The MarkedString.info is the former struct field name
*)

| EInj of expr Utils.Pos.marked * int * EnumName.t * typ Utils.Pos.marked list	(*
The MarkedString.info is the former enum case name
*)

| EMatch of expr Utils.Pos.marked * expr Utils.Pos.marked list * EnumName.t	(*
The MarkedString.info is the former enum case name
*)

| EArray of expr Utils.Pos.marked list

| ELit of lit

| EAbs of (expr, expr Utils.Pos.marked) Bindlib.mbinder Utils.Pos.marked * typ Utils.Pos.marked list

| EApp of expr Utils.Pos.marked * expr Utils.Pos.marked list

| EAssert of expr Utils.Pos.marked

| EOp of operator

| EDefault of expr Utils.Pos.marked list * expr Utils.Pos.marked * expr Utils.Pos.marked

| EIfThenElse of expr Utils.Pos.marked * expr Utils.Pos.marked * expr Utils.Pos.marked

| ErrorOnEmpty of expr Utils.Pos.marked

The expressions use the Bindlib library, based on higher-order abstract syntax

type struct_ctx = (StructFieldName.t * typ Utils.Pos.marked) list StructMap.t

type enum_ctx = (EnumConstructor.t * typ Utils.Pos.marked) list EnumMap.t

type decl_ctx = {

ctx_enums : enum_ctx;

ctx_structs : struct_ctx;

}

Variable helpers

module Var : sig ... end

module VarMap : Stdlib.Map.S with type key = Var.t

type vars = expr Bindlib.mvar

val make_var : Var.t Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box

val make_abs : vars -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> typ Utils.Pos.marked list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_app : expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_let_in : Var.t -> typ Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_multiple_let_in : Var.t array -> typ Utils.Pos.marked list -> expr Utils.Pos.marked Bindlib.box list -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

type binder = (expr, expr Utils.Pos.marked) Bindlib.binder

type program = {

decl_ctx : decl_ctx;

scopes : (ScopeName.t * Var.t * expr Utils.Pos.marked) list;

}