Module Ast.Infer

Contains some structures used for type inference

module Any : Utils.Uid.Id with type info = unit

type unionfind_typ = typ Utils.Marked.pos UnionFind.elem

We do not reuse typ because we have to include a new TAny variant. Indeed, error terms can have any type and this has to be captured by the type sytem.

and typ =

| TLit of typ_lit

| TArrow of unionfind_typ * unionfind_typ

| TTuple of unionfind_typ list * StructName.t option

| TEnum of unionfind_typ list * EnumName.t

| TArray of unionfind_typ

| TAny of Any.t

val typ_to_ast : unionfind_typ -> marked_typ

val ast_to_typ : marked_typ -> unionfind_typ