Module Dcalc.Autotest

This module scans a program for "test" scopes, detected as scopes without any undefined inputs. It runs the interpreter to compute their results, then inserts assertion in the code that ensure correctness of said results.

val program : 'm Ast.program -> 'm Ast.program