Module Dcalc.Interpreter

Reference interpreter for the default calculus

val empty_thunked_term : Ast.expr Utils.Pos.marked

val interpret_program : Ast.decl_ctx -> Ast.expr Utils.Pos.marked -> (Utils.Uid.MarkedString.info * Ast.expr Utils.Pos.marked) list

Interpret a program. This function expects an expression typed as a function whose argument are all thunked. The function is executed by providing for each argument a thunked empty default.