Module Dcalc__Interpreter

Reference interpreter for the default calculus

module Pos = Utils.Pos

module Errors = Utils.Errors

module Cli = Utils.Cli

module A = Dcalc.Ast

Helpers

val is_empty_error : A.expr Pos.marked -> bool

val empty_thunked_term : Dcalc.Ast.expr Pos.marked

val type_eq : A.typ Pos.marked -> A.typ Pos.marked -> bool

Evaluation

val evaluate_operator : A.operator Pos.marked -> A.expr Pos.marked list -> A.expr Pos.marked

val evaluate_expr : A.expr Pos.marked -> A.expr Pos.marked

API

val interpret_program : Dcalc.Ast.expr Pos.marked -> (Dcalc.Ast.Var.t * Dcalc.Ast.expr Pos.marked) list

Interpret a program. This function expects an expression typed as a function whose argument are all thunked. The function is executed by providing for each argument a thunked empty default.