Module Dcalc__Typing

Typing for the default calculus. Because of the error terms, we perform type inference using the classical W algorithm with union-find unification.

module Pos = Utils.Pos

module Errors = Utils.Errors

module A = Dcalc.Ast

module Cli = Utils.Cli

Types and unification

module Any : sig ... end

type typ =

| TLit of A.typ_lit

| TArrow of typ Pos.marked UnionFind.elem * typ Pos.marked UnionFind.elem

| TTuple of typ Pos.marked UnionFind.elem list * Dcalc.Ast.StructName.t option

| TEnum of typ Pos.marked UnionFind.elem list * Dcalc.Ast.EnumName.t

| TArray of typ Pos.marked UnionFind.elem

| TAny of Any.t

We do not reuse Ast.typ because we have to include a new TAny variant. Indeed, error terms can have any type and this has to be captured by the type sytem.

val typ_needs_parens : typ Pos.marked UnionFind.elem -> bool

val format_typ : Dcalc.Ast.decl_ctx -> Stdlib.Format.formatter -> typ Pos.marked UnionFind.elem -> unit

val unify : Dcalc.Ast.decl_ctx -> typ Pos.marked UnionFind.elem -> typ Pos.marked UnionFind.elem -> unit

Raises an error if unification cannot be performed

val op_type : A.operator Pos.marked -> typ Pos.marked UnionFind.elem

Operators have a single type, instead of being polymorphic with constraints. This allows us to have a simpler type system, while we argue the syntactic burden of operator annotations helps the programmer visualize the type flow in the code.

val ast_to_typ : A.typ -> typ

val typ_to_ast : typ Pos.marked UnionFind.elem -> A.typ Pos.marked

Double-directed typing

type env = typ Pos.marked UnionFind.elem A.VarMap.t

val typecheck_expr_bottom_up : Dcalc.Ast.decl_ctx -> env -> A.expr Pos.marked -> typ Pos.marked UnionFind.elem

Infers the most permissive type from an expression

val typecheck_expr_top_down : Dcalc.Ast.decl_ctx -> env -> A.expr A.Pos.marked -> typ Pos.marked UnionFind.elem -> unit

Checks whether the expression can be typed with the provided type

API

val infer_type : Dcalc.Ast.decl_ctx -> A.expr Pos.marked -> A.typ Pos.marked

val check_type : Dcalc.Ast.decl_ctx -> A.expr Pos.marked -> A.typ Pos.marked -> unit

Typechecks an expression given an expected type