Module Desugared

module Ast : sig ... end

Abstract syntax tree of the desugared representation

module Dependency : sig ... end

Scope dependencies computations using OCamlgraph

module Disambiguate : sig ... end

This module does local typing in order to fill some missing type information in the AST:

module From_surface : sig ... end

Translation from Surface.Ast to Desugared.Ast.

module Linting : sig ... end

module Name_resolution : sig ... end

Builds a context that allows for mapping each name to a precise uid, taking lexical scopes into account

module Print : sig ... end