Module Desugared__Ast

Abstract syntax tree of the desugared representation

Names, Maps and Keys

module IdentMap : Stdlib.Map.S with type IdentMap.key = Stdlib.String.t

module RuleName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module RuleMap : Stdlib.Map.S with type RuleMap.key = RuleName.t

module RuleSet : Stdlib.Set.S with type RuleSet.elt = RuleName.t

module LabelName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module LabelMap : Stdlib.Map.S with type LabelMap.key = LabelName.t

module LabelSet : Stdlib.Set.S with type LabelSet.elt = LabelName.t

module StateName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module ScopeVar : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module ScopeVarSet : Stdlib.Set.S with type ScopeVarSet.elt = ScopeVar.t

module ScopeVarMap : Stdlib.Map.S with type ScopeVarMap.key = ScopeVar.t

module ScopeDef : sig ... end

Inside a scope, a definition can refer either to a scope def, or a subscope def

module ScopeDefMap : Stdlib.Map.S with type ScopeDefMap.key = ScopeDef.t

module ScopeDefSet : Stdlib.Set.S with type ScopeDefSet.elt = ScopeDef.t

AST

type location =

| ScopeVar of ScopeVar.t Utils.Pos.marked * StateName.t option

| SubScopeVar of Scopelang.Ast.ScopeName.t * Scopelang.Ast.SubScopeName.t Utils.Pos.marked * ScopeVar.t Utils.Pos.marked

Expressions

module LocationSet : Stdlib.Set.S with type LocationSet.elt = location Utils.Pos.marked

type expr =

| ELocation of location

| EVar of expr Bindlib.var Utils.Pos.marked

| EStruct of Scopelang.Ast.StructName.t * expr Utils.Pos.marked Scopelang.Ast.StructFieldMap.t

| EStructAccess of expr Utils.Pos.marked * Scopelang.Ast.StructFieldName.t * Scopelang.Ast.StructName.t

| EEnumInj of expr Utils.Pos.marked * Scopelang.Ast.EnumConstructor.t * Scopelang.Ast.EnumName.t

| EMatch of expr Utils.Pos.marked * Scopelang.Ast.EnumName.t * expr Utils.Pos.marked Scopelang.Ast.EnumConstructorMap.t

| ELit of Dcalc.Ast.lit

| EAbs of (expr, expr Utils.Pos.marked) Bindlib.mbinder Utils.Pos.marked * Scopelang.Ast.typ Utils.Pos.marked list

| EApp of expr Utils.Pos.marked * expr Utils.Pos.marked list

| EOp of Dcalc.Ast.operator

| EDefault of expr Utils.Pos.marked list * expr Utils.Pos.marked * expr Utils.Pos.marked

| EIfThenElse of expr Utils.Pos.marked * expr Utils.Pos.marked * expr Utils.Pos.marked

| EArray of expr Utils.Pos.marked list

| ErrorOnEmpty of expr Utils.Pos.marked

The expressions use the Bindlib library, based on higher-order abstract syntax

Variable helpers

module Var : sig ... end

module VarMap : Stdlib.Map.S with type VarMap.key = Var.t

type vars = expr Bindlib.mvar

val make_var : Var.t Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box

val make_abs : vars -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> Scopelang.Ast.typ Utils.Pos.marked list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_app : expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_let_in : Var.t -> Scopelang.Ast.typ Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box

Rules and scopes

type rule = {

rule_id : RuleName.t;

rule_just : expr Utils.Pos.marked Bindlib.box;

rule_cons : expr Utils.Pos.marked Bindlib.box;

rule_parameter : (Var.t * Scopelang.Ast.typ Utils.Pos.marked) option;

rule_exception_to_rules : RuleSet.t Utils.Pos.marked;

}

val empty_rule : Utils.Pos.t -> Scopelang.Ast.typ Utils.Pos.marked option -> rule

val always_false_rule : Utils.Pos.t -> Scopelang.Ast.typ Utils.Pos.marked option -> rule

type assertion = expr Utils.Pos.marked Bindlib.box

type variation_typ =

| Increasing

| Decreasing

type reference_typ =

| Decree

| Law

type meta_assertion =

| FixedBy of reference_typ Utils.Pos.marked

| VariesWith of unit * variation_typ Utils.Pos.marked option

type scope_def = {

scope_def_rules : rule RuleMap.t;

scope_def_typ : Scopelang.Ast.typ Utils.Pos.marked;

scope_def_is_condition : bool;

scope_def_io : Scopelang.Ast.io;

scope_def_label_groups : RuleSet.t LabelMap.t;

}

type var_or_states =

| WholeVar

| States of StateName.t list

type scope = {

scope_vars : var_or_states ScopeVarMap.t;

scope_sub_scopes : Scopelang.Ast.ScopeName.t Scopelang.Ast.SubScopeMap.t;

scope_uid : Scopelang.Ast.ScopeName.t;

scope_defs : scope_def ScopeDefMap.t;

scope_assertions : assertion list;

scope_meta_assertions : meta_assertion list;

}

type program = {

program_scopes : scope Scopelang.Ast.ScopeMap.t;

program_enums : Scopelang.Ast.enum_ctx;

program_structs : Scopelang.Ast.struct_ctx;

}

Helpers

val locations_used : expr Utils.Pos.marked -> LocationSet.t

val free_variables : rule RuleMap.t -> Utils.Pos.t ScopeDefMap.t