Module Desugared__Desugared_to_scope

Translation from Ast to Ast

module Pos = Utils.Pos

module Errors = Utils.Errors

module Cli = Utils.Cli

Rule tree construction

type rule_tree =

| Leaf of Desugared.Ast.rule

| Node of rule_tree list * Desugared.Ast.rule

val def_map_to_tree : Desugared.Ast.ScopeDef.t -> Scopelang.Ast.typ Pos.marked option -> Desugared.Ast.rule Desugared.Ast.RuleMap.t -> rule_tree list

Transforms a flat list of rules into a tree, taking into account the priorities declared between rules

Invariant: there are no exceptions cycles

Invariant: there are no dandling exception pointers in the rules

val rule_tree_to_expr : toplevel:bool -> Pos.t -> Scopelang.Ast.Var.t option -> rule_tree -> Scopelang.Ast.expr Pos.marked Bindlib.box

From the rule_tree, builds an Ast.expr.EDefault expression in the scope language. The ~toplevel parameter is used to know when to place the toplevel binding in the case of functions.

AST translation

val translate_def : Desugared.Ast.ScopeDef.t -> Desugared.Ast.rule Desugared.Ast.RuleMap.t -> Scopelang.Ast.typ Pos.marked -> bool -> Scopelang.Ast.expr Pos.marked

Translates a definition inside a scope, the resulting expression should be an Ast.expr.EDefault

val translate_scope : Desugared.Ast.scope -> Scopelang.Ast.scope_decl

Translates a scope

API

val translate_program : Desugared.Ast.program -> Scopelang.Ast.program