Module Json_schema

Catala plugin for generating JSON schemas used to build forms for the Catala website.

val name : string

val extension : string

module D = Dcalc.Ast

module To_json : sig ... end

Contains all format functions used to format a Lcalc Catala program representation to a JSON schema describing the corresponding web form.

val apply : source_file:Catala_utils.Pos.input_file -> output_file:string option -> scope:Shared_ast.ScopeName.t option -> 'm Lcalc.Ast.program -> Scopelang.Dependency.TVertex.t list -> unit