Module Lcalc

module Ast : sig ... end

Abstract syntax tree for the lambda calculus

module Closure_conversion : sig ... end

This module performs environment-passing style closure conversion, relying on the existential TClosureEnv type and tuples for closure environments. The implementation is based on François Pottier's MPRI lesson. After closure conversion, closure hoisting is perform and all closures end up as toplevel definitions.

module From_dcalc : sig ... end

Translation from the default calculus to the lambda calculus.

module Lift_positions : sig ... end

This modules does the following:

module Monomorphize : sig ... end

module To_ocaml : sig ... end