Module Lcalc.Ast

Abstract syntax tree for the lambda calculus

Abstract syntax tree

The expressions use the Bindlib library, based on higher-order abstract syntax

type lit =

| LBool of bool

| LInt of Runtime.integer

| LRat of Runtime.decimal

| LMoney of Runtime.money

| LUnit

| LDate of Runtime.date

| LDuration of Runtime.duration

type except =

| ConflictError

| EmptyError

| NoValueProvided

| Crash

type expr =

| EVar of expr Bindlib.var Utils.Pos.marked

| ETuple of expr Utils.Pos.marked list * Dcalc.Ast.StructName.t option	(*
The MarkedString.info is the former struct field name
*)

| ETupleAccess of expr Utils.Pos.marked * int * Dcalc.Ast.StructName.t option * Dcalc.Ast.typ Utils.Pos.marked list	(*
The MarkedString.info is the former struct field name
*)

| EInj of expr Utils.Pos.marked * int * Dcalc.Ast.EnumName.t * Dcalc.Ast.typ Utils.Pos.marked list	(*
The MarkedString.info is the former enum case name
*)

| EMatch of expr Utils.Pos.marked * expr Utils.Pos.marked list * Dcalc.Ast.EnumName.t	(*
The MarkedString.info is the former enum case name
*)

| EArray of expr Utils.Pos.marked list

| ELit of lit

| EAbs of ( expr, expr Utils.Pos.marked ) Bindlib.mbinder Utils.Pos.marked * Dcalc.Ast.typ Utils.Pos.marked list

| EApp of expr Utils.Pos.marked * expr Utils.Pos.marked list

| EAssert of expr Utils.Pos.marked

| EOp of Dcalc.Ast.operator

| EIfThenElse of expr Utils.Pos.marked * expr Utils.Pos.marked * expr Utils.Pos.marked

| ERaise of except

| ECatch of expr Utils.Pos.marked * except * expr Utils.Pos.marked

type scope_body = {

scope_body_name : Dcalc.Ast.ScopeName.t;

scope_body_var : expr Bindlib.var;

scope_body_expr : expr Utils.Pos.marked;

}

type program = {

decl_ctx : Dcalc.Ast.decl_ctx;

scopes : scope_body list;

}

Variable helpers

module Var : sig ... end

module VarMap : Stdlib.Map.S with type key = Var.t

module VarSet : Stdlib.Set.S with type elt = Var.t

type vars = expr Bindlib.mvar

type binder = ( expr, expr Utils.Pos.marked ) Bindlib.binder

Language terms construction

val make_var : Var.t Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box

val make_abs : vars -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> Dcalc.Ast.typ Utils.Pos.marked list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_app : expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_let_in : Var.t -> Dcalc.Ast.typ Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box

val make_multiple_let_in : Var.t array -> Dcalc.Ast.typ Utils.Pos.marked list -> expr Utils.Pos.marked Bindlib.box list -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val option_enum : Dcalc.Ast.EnumName.t

val none_constr : Dcalc.Ast.EnumConstructor.t

val some_constr : Dcalc.Ast.EnumConstructor.t

val option_enum_config : (Dcalc.Ast.EnumConstructor.t * Dcalc.Ast.typ Utils.Pos.marked) list

val make_none : Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_some : expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box

val make_matchopt_with_abs_arms : expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box

val make_matchopt : Utils.Pos.t -> Var.t -> Dcalc.Ast.typ Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box

e' = make_matchopt'' pos v e e_none e_some Builds the term corresponding to match e with | None -> fun () -> e_none |Some -> fun v -> e_some.

Special symbols

val handle_default : Var.t

val handle_default_opt : Var.t