Module Lcalc.Closure_conversion

This module performs environment-passing style closure conversion, relying on the existential TClosureEnv type and tuples for closure environments. The implementation is based on François Pottier's MPRI lesson. After closure conversion, closure hoisting is perform and all closures end up as toplevel definitions.

val closure_conversion : keep_special_ops:bool -> 'm Ast.program -> 'm Ast.program