Module Lcalc.Compile_without_exceptions

Translation from the default calculus to the lambda calculus. This translation uses an option monad to handle empty defaults terms. This transformation is one piece to permit to compile toward legacy languages that does not contains exceptions.

val translate_program : Shared_ast.typed Dcalc.Ast.program -> Shared_ast.untyped Ast.program