Module Literate.Literate_common

val literal_title : Catala_utils.Global.backend_lang -> string

Return the title traduction according the given Catala_utils.Global.backend_lang.

val literal_generated_by : Catala_utils.Global.backend_lang -> string

Return the 'generated by' traduction according the given Catala_utils.Global.backend_lang.

val literal_source_files : Catala_utils.Global.backend_lang -> string

Return the 'source files weaved' traduction according the given Catala_utils.Global.backend_lang.

val literal_disclaimer_and_link : Catala_utils.Global.backend_lang -> string

Return the traduction of a paragraph giving a basic disclaimer about Catala and a link to the website according the given Catala_utils.Global.backend_lang.

val literal_last_modification : Catala_utils.Global.backend_lang -> string

Return the 'last modification' traduction according the given Catala_utils.Global.backend_lang.

val get_language_extension : Catala_utils.Global.backend_lang -> string

Return the file extension corresponding to the given Catala_utils.Global.backend_lang.

val run_pandoc : string -> [ `Html | `Latex ] -> string

Runs the pandoc on a string to pretty-print markdown features into the desired format.

val check_exceeding_lines : ?max_len:int -> int -> string -> string -> unit

check_exceeding_lines ~max_len start_line filename content prints a warning message for each lines of content exceeding max_len characters.

val call_pygmentize : ?lang:Catala_utils.Global.backend_lang -> string list -> string

Calls the pygmentize command with the given arguments, and returns the results as a string. If lang is specified, the proper arguments for the Catala lexer are already passed.

val with_pygmentize_lexer : Catala_utils.Global.backend_lang -> (string list -> 'a) -> 'a

Creates the required lexer file and returns the corresponding pygmentize command-line arguments