Module Python

This file is only for demonstration purposes, showing a trivial use of backend plugins for Catala.

The code for the Python backend already has first-class support, so there would be no reason to use this plugin instead

val name : string

val extension : string

val apply : source_file:'a -> output_file:string option -> scope:'b -> Scalc.Ast.program -> Scopelang.Dependency.TVertex.t list -> unit