Class type Runtime.raw_event

Wrapper for the Runtime_ocaml.Runtime.raw_event -- directly collected during the program execution.

method eventType : Js_of_ocaml.Js.js_string Js_of_ocaml.Js.t Js_of_ocaml.Js.prop

There is four type of raw log events:

'BeginCall' is emitted when a function or a subscope is called.

'EndCall' is emitted when a function or a subscope is exited.

'VariableDefinition' is emitted when a variable or a function is defined.

'DecisionTaken' stores the information about the source position of the event.

method information : Js_of_ocaml.Js.js_string Js_of_ocaml.Js.t Js_of_ocaml.Js.js_array Js_of_ocaml.Js.t Js_of_ocaml.Js.prop

Represents information about a name in the code -- i.e. variable name, subscope name, etc...

It's a list of strings with a length varying from 2 to 3, where:

the first string is the name of the current scope -- starting with a capitalized letter Scope_name,

the second string is either: the name of a scope variable or, the name of a subscope input variable -- a_subscope_var.input_var

the third string is either: a subscope name (starting with a capitalized letter Subscope_name or, the input (resp. output) string -- which corresponds to the input (resp. the output) of a function.

method sourcePosition : source_position Js_of_ocaml.Js.t Js_of_ocaml.Js.optdef Js_of_ocaml.Js.prop

method loggedValueJson : Js_of_ocaml.Js.js_string Js_of_ocaml.Js.t Js_of_ocaml.Js.prop

Serialized Runtime_ocaml.Runtime.runtime_value corresponding to a 'VariableDefinition' raw event.