Module Scalc

module Ast : sig ... end

module From_lcalc : sig ... end

module Print : sig ... end

module To_c : sig ... end

Formats a statement calculus program into a valid C89 program

module To_java : sig ... end

Formats a lambda calculus program into a valid Java program

module To_python : sig ... end

Formats a lambda calculus program into a valid Python program

module Utils : sig ... end