Module Scopelang.Ast

Abstract syntax tree of the scope language

Identifiers

module ScopeMap : Stdlib.Map.S with type key = Shared_ast.ScopeName.t

module SubScopeNameSet : Stdlib.Set.S with type elt = Shared_ast.SubScopeName.t

module SubScopeMap : Stdlib.Map.S with type key = Shared_ast.SubScopeName.t

module StructFieldMapLift : sig ... end

module EnumConstructorMapLift : sig ... end

type location = Shared_ast.scopelang Shared_ast.glocation

module LocationSet : Stdlib.Set.S with type elt = location Utils.Marked.pos

Abstract syntax tree

type expr = ( Shared_ast.scopelang, Shared_ast.untyped Shared_ast.mark ) Shared_ast.gexpr

module ExprMap : Stdlib.Map.S with type key = expr

val locations_used : expr -> LocationSet.t

type io_input =

| NoInput	(*
For an internal variable defined only in the scope, and does not appear in the input.
*)

| OnlyInput	(*
For variables that should not be redefined in the scope, because they appear in the input.
*)

| Reentrant	(*
For variables defined in the scope that can also be redefined by the caller as they appear in the input.
*)

This type characterizes the three levels of visibility for a given scope variable with regards to the scope's input and possible redefinitions inside the scope..

type io = {

io_output : bool Utils.Marked.pos;	(*
true is present in the output of the scope.
*)

io_input : io_input Utils.Marked.pos;

}

Characterization of the input/output status of a scope variable.

type rule =

| Definition of location Utils.Marked.pos * Shared_ast.typ * io * expr

| Assertion of expr

| Call of Shared_ast.ScopeName.t * Shared_ast.SubScopeName.t

type scope_decl = {

scope_decl_name : Shared_ast.ScopeName.t;

scope_sig : (Shared_ast.typ * io) Shared_ast.ScopeVarMap.t;

scope_decl_rules : rule list;

}

type program = {

program_scopes : scope_decl ScopeMap.t;

program_ctx : Shared_ast.decl_ctx;

}