Module Shared_ast

This module defines generic types for types, literals and expressions shared through several of the different ASTs.

module ModuleName = Catala_utils.Uid.Module

module ScopeName : sig ... end

module TopdefName : sig ... end

module StructName : sig ... end

module StructField : sig ... end

module EnumName : sig ... end

module EnumConstructor : sig ... end

module AbstractType : sig ... end

Only used by surface

module RuleName : sig ... end

module LabelName : sig ... end

Used for unresolved structs/maps in desugared

module Ident = Catala_utils.String

module MarkedIdent = Catala_utils.Uid.MarkedString

Only used by desugared/scopelang

module ScopeVar : sig ... end

type scope_var_or_subscope =

| ScopeVar of ScopeVar.t

| SubScope of ScopeVar.t * ScopeName.t

module StateName : sig ... end

module ConstantNames : sig ... end

Abstract syntax tree

Define a common base type for the expressions in most passes of the compiler

Phantom types used to select relevant cases on the generic AST

we instantiate them with a polymorphic variant to take advantage of sub-typing. The values aren't actually used.

These types allow to select the features present in any given expression type

type yes =

| Yes

type no =

| No
(*
Phantom types used in the definitions below. We don't make them abstract, because the typer needs to know that their intersection is empty.
*)

type desugared = < monomorphic : yes ; polymorphic : yes ; overloaded : yes ; resolved : no ; syntacticNames : yes ; scopeVarStates : yes ; scopeVarSimpl : no ; explicitScopes : yes ; defaultTerms : yes ; reifiedPos : no ; custom : no >

type scopelang = < monomorphic : yes ; polymorphic : yes ; overloaded : no ; resolved : yes ; syntacticNames : no ; scopeVarStates : no ; scopeVarSimpl : yes ; explicitScopes : yes ; defaultTerms : yes ; reifiedPos : no ; custom : no >

type dcalc = < monomorphic : yes ; polymorphic : yes ; overloaded : no ; resolved : yes ; syntacticNames : no ; scopeVarStates : no ; scopeVarSimpl : no ; explicitScopes : no ; defaultTerms : yes ; reifiedPos : no ; custom : no >

type lcalc = < monomorphic : yes ; polymorphic : yes ; overloaded : no ; resolved : yes ; syntacticNames : no ; scopeVarStates : no ; scopeVarSimpl : no ; explicitScopes : no ; defaultTerms : no ; reifiedPos : yes ; custom : no >

type 'a any = < .. > as 'a

'a any is 'a, but adds the constraint that it should be restricted to valid AST kinds

type dcalc_lcalc_features = < monomorphic : yes ; polymorphic : yes ; overloaded : no ; resolved : yes ; syntacticNames : no ; scopeVarStates : no ; scopeVarSimpl : no ; explicitScopes : no >

Features that are common to Dcalc and Lcalc

type ('d, 'r) dcalc_lcalc = < dcalc_lcalc_features ; defaultTerms : 'd ; reifiedPos : 'r ; custom : no >

This type regroups Dcalc and Lcalc ASTs.

type ('d, 'r, 'c) interpr_kind = < dcalc_lcalc_features ; defaultTerms : 'd ; reifiedPos : 'r ; custom : 'c >

This type corresponds to the types handled by the interpreter: it regroups Dcalc and Lcalc ASTs and may have custom terms

Types

type typ_lit =

| TUnit

| TBool

| TInt

| TMoney

| TRat

| TDate

| TDuration

| TPos

type typ = naked_typ Catala_utils.Mark.pos

and naked_typ =

| TError

| TLit of typ_lit

| TArrow of typ list * typ

| TTuple of typ list

| TStruct of StructName.t

| TEnum of EnumName.t

| TAbstract of AbstractType.t

| TOption of typ

| TArray of typ

| TDefault of typ

| TVar of naked_typ Bindlib.var

| TForAll of (naked_typ, typ) Bindlib.mbinder
(*
Universal quantification of type variables
*)

| TClosureEnv
(*
Hides an existential type needed for closure conversion
*)

module TypeIdent : sig ... end

Constants and operators

type date = Catala_runtime.date

type date_rounding = Catala_runtime.date_rounding

type duration = Catala_runtime.duration

type var_def_log = {

log_typ : naked_typ;

log_io_input : Catala_runtime.io_input;

log_io_output : bool;

}

type log_entry =

| VarDef of var_def_log
(*
During code generation, we need to know the type of the variable being logged for embedding as well as its I/O properties.
*)

| BeginCall

| EndCall

| PosRecordIfTrueBool

module Op : sig ... end

Classification of operators on how they should be typed

type 'a operator = 'a Op.t

type attr_path = string list Catala_utils.Mark.pos

Position-embedded attributes

type attr_value = ..

type attr_value +=

| String of string Catala_utils.Mark.pos

| Unit

type Catala_utils.Pos.attr +=

| Src of attr_path * attr_value * Catala_utils.Pos.t

| DebugPrint of {

label : string option;

}

| Test

| Doc of string * Catala_utils.Pos.t

| ImplicitPosArg

| ErrorMessage of string

| JsonPayload of string

Markings

type untyped = {

pos : Catala_utils.Pos.t;

}

type typed = {

pos : Catala_utils.Pos.t;

ty : typ;

}

type 'a custom = {

pos : Catala_utils.Pos.t;

custom : 'a;

}

type nil = |

Using empty markings will ensure terms can't be constructed: used for example in interfaces to ensure that they don't contain any expressions

type _ mark =

| Untyped : untyped -> untyped mark

| Typed : typed -> typed mark

| Custom : 'a custom -> 'a custom mark

The generic type of AST markings. Using a GADT allows functions to be polymorphic in the marking, but still do transformations on types when appropriate. The Custom case can be used within passes that need to store specific information, e.g. typing

type ('a, 'm) marked = ('a, 'm mark) Catala_utils.Mark.ed

Type of values marked with the above standard mark GADT

Generic expressions

Define a common base type for the expressions in most passes of the compiler

type lit =

| LBool of bool

| LInt of Catala_runtime.integer

| LRat of Catala_runtime.decimal

| LMoney of Catala_runtime.money

| LUnit

| LDate of date

| LDuration of duration

Literals are the same throughout compilation.

type external_ref =

| External_value of TopdefName.t

| External_scope of ScopeName.t

External references are resolved to strings that point to functions or constants in the end, but we need to keep different references for typing

type 'a glocation =

| DesugaredScopeVar : {

name : ScopeVar.t Catala_utils.Mark.pos;

state : StateName.t option;

} -> < scopeVarStates : yes.. > glocation

| ScopelangScopeVar : {

name : ScopeVar.t Catala_utils.Mark.pos;

} -> < scopeVarSimpl : yes.. > glocation

| ToplevelVar : {

name : TopdefName.t Catala_utils.Mark.pos;

is_external : bool;

} -> < explicitScopes : yes.. > glocation

Locations are handled differently in desugared and scopelang

type ('a, 'm) gexpr = (('a, 'm) naked_gexpr, 'm) marked

and ('a, 'm) naked_gexpr = ('a, 'a, 'm) base_gexpr

General expressions: groups all expression cases of the different ASTs, and uses a GADT to eliminate irrelevant cases for each one. The 't annotations are also totally unconstrained at this point. The dcalc exprs, for ex ample, are then defined with type naked_expr = dcalc naked_gexpr plus the annotations.

A few tips on using this GADT:

To write a function that handles cases from different ASTs, explicit the type variables: fun (type a) (x: a naked_gexpr) -> ...

For recursive functions, you may need to additionally explicit the generalisation of the variable: let rec f: type a . a naked_gexpr -> ...

Always think of using the pre-defined map/fold functions in Expr rather than completely defining your recursion manually.

The first argument of the base_gexpr type caracterises the "deep" type of the AST, while the second is the shallow type. They are always equal for well-formed AST types, but differentiating them ephemerally allows us to do well-typed recursive transformations on the AST that change its type

and ('a, 'b, 'm) base_gexpr =

| ELit : lit -> ('a, < .. >, 'm) base_gexpr

| EApp : {

f : ('a, 'm) gexpr;

args : ('a, 'm) gexpr list;
(*
length may be 1 even if arity > 1 in desugared. scopelang performs detuplification, so length = arity afterwards
*)

tys : typ list;
(*
Set to [] before disambiguation
*)

} -> ('a, < .. >, 'm) base_gexpr

| EAppOp : {

op : 'a operator Catala_utils.Mark.pos;

args : ('a, 'm) gexpr list;

tys : typ list;

} -> ('a, < .. >, 'm) base_gexpr

| EArray : ('a, 'm) gexpr list -> ('a, < .. >, 'm) base_gexpr

| EVar : ('a, 'm) naked_gexpr Bindlib.var -> ('a, _, 'm) base_gexpr

| EAbs : {

binder : (('a, 'a, 'm) base_gexpr, ('a, 'm) gexpr) Bindlib.mbinder;

pos : Catala_utils.Pos.t list;

tys : typ list;

} -> ('a, < .. >, 'm) base_gexpr

| EIfThenElse : {

cond : ('a, 'm) gexpr;

etrue : ('a, 'm) gexpr;

efalse : ('a, 'm) gexpr;

} -> ('a, < .. >, 'm) base_gexpr

| EStruct : {

name : StructName.t;

fields : ('a, 'm) gexpr StructField.Map.t;

} -> ('a, < .. >, 'm) base_gexpr

| EInj : {

name : EnumName.t;

e : ('a, 'm) gexpr;

cons : EnumConstructor.t;

} -> ('a, < .. >, 'm) base_gexpr

| EMatch : {

name : EnumName.t;

e : ('a, 'm) gexpr;

cases : ('a, 'm) gexpr EnumConstructor.Map.t;

} -> ('a, < .. >, 'm) base_gexpr

| ETuple : ('a, 'm) gexpr list -> ('a, < .. >, 'm) base_gexpr

| ETupleAccess : {

e : ('a, 'm) gexpr;

index : int;

size : int;

} -> ('a, < .. >, 'm) base_gexpr

| EAssert : ('a, 'm) gexpr -> ('a, < reifiedPos : no.. >, 'm) base_gexpr

| ELocation : 'b glocation -> ('a, < .. > as 'b, 'm) base_gexpr

| EScopeCall : {

scope : ScopeName.t;

args : (Catala_utils.Pos.t * ('a, 'm) gexpr) ScopeVar.Map.t;

} -> ('a, < explicitScopes : yes.. >, 'm) base_gexpr

| EDStructAmend : {

name_opt : StructName.t option;

e : ('a, 'm) gexpr;

fields : ('a, 'm) gexpr MarkedIdent.Map.t;

} -> ('a, < syntacticNames : yes.. >, 'm) base_gexpr

| EDStructAccess : {

name_opt : StructName.t option;

e : ('a, 'm) gexpr;

field : MarkedIdent.t;

} -> ('a, < syntacticNames : yes.. >, 'm) base_gexpr
(*
desugared has ambiguous struct fields
*)

| EStructAccess : {

name : StructName.t;

e : ('a, 'm) gexpr;

field : StructField.t;

} -> ('a, < .. >, 'm) base_gexpr
(*
Resolved struct/enums, after name resolution in desugared
*)

| EExternal : {

name : external_ref Catala_utils.Mark.pos;

} -> ('a, < explicitScopes : no.. >, 't) base_gexpr

| EFatalError : Catala_runtime.error -> ('a, < reifiedPos : no.. >, 'm) base_gexpr

| EFatalError_pos : {

error : Catala_runtime.error;

pos_expr : ('a, 'm) gexpr;

} -> ('a, < reifiedPos : yes.. >, 'm) base_gexpr

| EPos : Catala_utils.Pos.t -> ('a, < .. >, 'm) base_gexpr

| EDefault : {

excepts : ('a, 'm) gexpr list;

just : ('a, 'm) gexpr;

cons : ('a, 'm) gexpr;

} -> ('a, < defaultTerms : yes.. >, 'm) base_gexpr

| EPureDefault : ('a, 'm) gexpr -> ('a, < defaultTerms : yes.. >, 'm) base_gexpr
(*
"return" of a pure term, so that it can be typed as default
*)

| EEmpty : ('a, < defaultTerms : yes.. >, 'm) base_gexpr

| EErrorOnEmpty : ('a, 'm) gexpr -> ('a, < defaultTerms : yes.. >, 'm) base_gexpr

| ECustom : {

obj : Stdlib.Obj.t;

targs : typ list;

tret : typ;

} -> ('a, < custom : yes.. >, 't) base_gexpr
(*
A value or function of the given type, as a runtime OCaml object. For values, targs must be the empty list, and tret is the actual type. The specified types for arguments and result must be the Catala types corresponding to the runtime types .
*)

| EBad : ('a, < .. >, 't) base_gexpr
(*
An internal-only node used to progress through the passes to report multiple errors. It should not appear to the user.
*)

type any_expr =

| AnyExpr : ('a, _) gexpr -> any_expr

Useful for errors and printing, for example

type ('a, 'm) boxed_gexpr = (('a, 'm) naked_gexpr Bindlib.box, 'm) marked

The annotation is lifted outside of the box for expressions

type 'e boxed = ('a, 'm) boxed_gexpr constraint 'e = ('a, 'm) gexpr

('a, 'm) gexpr boxed is ('a, 'm) boxed_gexpr. The difference with ('a, 'm) gexpr Bindlib.box is that the annotations is outside of the box, and can therefore be accessed without the need to resolve the box

type ('e, 'b) binder = (('a, 'm) naked_gexpr, 'b) Bindlib.binder constraint 'e = ('a, 'm) gexpr

The expressions use the Bindlib library, based on higher-order abstract syntax

type ('e, 'b) mbinder = (('a, 'm) naked_gexpr, 'b) Bindlib.mbinder constraint 'e = ('a, 'm) gexpr

Higher-level program structure

Constructs scopes and programs on top of expressions. The 'e type parameter throughout is expected to match instances of the gexpr type defined above. Markings are constrained to the mark GADT defined above. Note that this structure is at the moment only relevant for dcalc and lcalc, as scopelang has its own scope structure, as the name implies.

type ('e, 'elt, 'last) bound_list =

| Last of 'last

| Cons of 'elt * ('e, ('e, 'elt, 'last) bound_list) binder

A linked list, but with a binder for each element into the next: x := let a = e1 in e2 is thus Cons (e1, {a. Cons (e2, {x. Nil})})

type scope_let_kind =

| DestructuringInputStruct
(*
let x = input.field
*)

| ScopeVarDefinition
(*
let x = error_on_empty e
*)

| CallingSubScope
(*
let result = s ({ x = s.x; y = s.x; ...})
*)

| DestructuringSubScopeResults
(*
let s.x = result.x *
*)

| Assertion
(*
let () = assert e
*)

This kind annotation signals that the let-binding respects a structural invariant. These invariants concern the shape of the expression in the let-binding, and are documented below.

type 'e scope_let = {

scope_let_kind : scope_let_kind;

scope_let_typ : typ;

scope_let_expr : 'e;

scope_let_pos : Catala_utils.Pos.t;

} constraint 'e = ('a any, _) gexpr

This type is parametrized by the expression type so it can be reused in later intermediate representations.

type visibility =

| Private

| Public

type 'e scope_body_expr = ('e, 'e scope_let, 'e) bound_list constraint 'e = ('a any, _) gexpr

A scope let-binding has all the information necessary to make a proper let-binding expression, plus an annotation for the kind of the let-binding that comes from the compilation of a Scopelang.Ast statement.

type 'e scope_body = {

scope_body_input_struct : StructName.t;

scope_body_output_struct : StructName.t;

scope_body_expr : ('e, 'e scope_body_expr) binder;

scope_body_visibility : visibility;

} constraint 'e = ('a any, _) gexpr

Instead of being a single expression, we give a little more ad-hoc structure to the scope body by decomposing it in an ordered list of let-bindings, and a result expression that uses the let-binded variables. The first binder is the argument of type scope_body_input_struct.

type 'e code_item =

| ScopeDef of ScopeName.t * 'e scope_body

| Topdef of TopdefName.t * typ * visibility * 'e

type code_export_kind =

| KScope of ScopeName.t

| KTopdef of TopdefName.t

| KTest of ScopeName.t

type 'e code_export = code_export_kind * 'e constraint 'e = ('a any, _) gexpr

Contains the code for the call to registered test scopes (useful e.g. for crafting a "main" function)

type 'e code_item_list = ('e, 'e code_item, 'e code_export list) bound_list

type struct_ctx = typ StructField.Map.t StructName.Map.t

type enum_ctx = typ EnumConstructor.Map.t EnumName.Map.t

type scope_info = {

in_struct_name : StructName.t;

out_struct_name : StructName.t;

out_struct_fields : StructField.t ScopeVar.Map.t;

visibility : visibility;

}

type module_intf_id = {

hash : Catala_utils.Hash.t;

is_stdlib : bool;

is_external : bool;

}

type module_tree_node = {

deps : module_tree;

intf_id : module_intf_id;

}

and module_tree = module_tree_node ModuleName.Map.t

In practice, this is a DAG: beware of repeated names

type decl_ctx = {

ctx_enums : enum_ctx;

ctx_structs : struct_ctx;

ctx_abstract_types : AbstractType.Set.t;

ctx_scopes : scope_info ScopeName.Map.t;

ctx_topdefs : (typ * visibility) TopdefName.Map.t;

ctx_public_types : TypeIdent.Set.t;
(*
types that need to be exported in the backend interfaces
*)

ctx_struct_fields : StructField.t StructName.Map.t Ident.Map.t;
(*
needed for disambiguation (desugared -> scope)
*)

ctx_enum_constrs : EnumConstructor.t EnumName.Map.t Ident.Map.t;

ctx_scope_index : ScopeName.t Ident.Map.t;
(*
only used to lookup scopes (in the root module) specified from the cli
*)

ctx_modules : module_tree;

}

type 'e program = {

decl_ctx : decl_ctx;

code_items : 'e code_item_list;

lang : Catala_utils.Global.backend_lang;

module_name : (ModuleName.t * module_intf_id) option;

}

module Var : sig ... end

module Qident : sig ... end

This module defines module names and path accesses, used to refer to separate compilation units.

module Type : sig ... end

module Operator : sig ... end

module Expr : sig ... end

Functions handling the expressions of shared_ast

module BoundList : sig ... end

Bound lists are non-empty linked lists where each element is a binder onto the next. They are useful for ordered program definitions, like nested let-ins.

module Scope : sig ... end

Functions handling the code item structures of shared_ast, in particular the scopes

module Program : sig ... end

module Renaming : sig ... end

module Print : sig ... end

Printing functions for the default calculus AST

module Typing : sig ... end

Typing for the default calculus. Because of the error terms, we perform type inference using the classical W algorithm with union-find unification.

module Interpreter : sig ... end

Reference interpreter for the default calculus

module Optimizations : sig ... end

Optimization passes for default calculus and lambda calculus programs and expressions

module Coverage : sig ... end

module Encoding : sig ... end

JSON encoding functions for Catala values