Module Shared_ast.Op

Classification of operators on how they should be typed

type monomorphic = < monomorphic : Shared_ast__.Definitions.yes >

Operands and return types of the operator are fixed

type polymorphic = < polymorphic : Shared_ast__.Definitions.yes >

The operator is truly polymorphic: it's the same runtime function that may work on multiple types. We require that resolving the argument types from right to left trivially resolves all type variables declared in the operator type.

type overloaded = < overloaded : Shared_ast__.Definitions.yes >

The operator is ambiguous and requires the types of its arguments to be known before it can be typed, using a pre-defined table

type resolved = < resolved : Shared_ast__.Definitions.yes >

Explicit monomorphic versions of the overloaded operators

type _ t =

| Not : < monomorphic.. > t

| ValueFromJson : Shared_ast__.Definitions.naked_typ Catala_utils.Mark.pos * string -> < monomorphic.. > t

| Length : < polymorphic.. > t

| Log : Shared_ast__.Definitions.log_entry * Catala_utils.Uid.MarkedString.info list -> < polymorphic.. > t

| ToClosureEnv : < polymorphic.. > t

| FromClosureEnv : < polymorphic.. > t

| ArrayAccess : int -> < polymorphic.. > t

| ConstructorCheck : (Shared_ast__.Definitions.EnumName.t * Shared_ast__.Definitions.EnumConstructor.t) -> < polymorphic.. > t

| DebugPrint : string -> < polymorphic.. > t

| Minus : < overloaded.. > t

| Minus_int : < resolved.. > t

| Minus_rat : < resolved.. > t

| Minus_mon : < resolved.. > t

| Minus_dur : < resolved.. > t

| ToInt : < overloaded.. > t

| ToInt_rat : < resolved.. > t

| ToInt_mon : < resolved.. > t

| ToRat : < overloaded.. > t

| ToRat_int : < resolved.. > t

| ToRat_mon : < resolved.. > t

| ToMoney : < overloaded.. > t

| ToMoney_rat : < resolved.. > t

| ToMoney_int : < resolved.. > t

| Round : < overloaded.. > t

| Round_rat : < resolved.. > t

| Round_mon : < resolved.. > t

| And : < monomorphic.. > t

| Or : < monomorphic.. > t

| Xor : < monomorphic.. > t

| Eq : < polymorphic.. > t

| Lt : < polymorphic.. > t

| Lte : < polymorphic.. > t

| Gt : < polymorphic.. > t

| Gte : < polymorphic.. > t

| Concat : < polymorphic.. > t

| Map : < polymorphic.. > t

| Filter : < polymorphic.. > t

| Find : < polymorphic.. > t

| Reduce : < polymorphic.. > t

| Sort : [ `Asc | `Desc ] -> < polymorphic.. > t

| Add : < overloaded.. > t

| Add_int_int : < resolved.. > t

| Add_rat_rat : < resolved.. > t

| Add_mon_mon : < resolved.. > t

| Add_dat_dur : Catala_runtime.date_rounding -> < resolved.. > t

| Add_dur_dur : < resolved.. > t

| Sub : < overloaded.. > t

| Sub_int_int : < resolved.. > t

| Sub_rat_rat : < resolved.. > t

| Sub_mon_mon : < resolved.. > t

| Sub_dat_dat : < resolved.. > t

| Sub_dat_dur : Catala_runtime.date_rounding -> < resolved.. > t

| Sub_dur_dur : < resolved.. > t

| Mult : < overloaded.. > t

| Mult_int_int : < resolved.. > t

| Mult_rat_rat : < resolved.. > t

| Mult_mon_int : < resolved.. > t

| Mult_mon_rat : < resolved.. > t

| Mult_dur_int : < resolved.. > t

| Div : < overloaded.. > t

| Div_int_int : < resolved.. > t

| Div_rat_rat : < resolved.. > t

| Div_mon_int : < resolved.. > t

| Div_mon_rat : < resolved.. > t

| Div_mon_mon : < resolved.. > t

| Div_dur_dur : < resolved.. > t

| Map2 : < polymorphic.. > t

| Fold : < polymorphic.. > t

| HandleExceptions : < polymorphic.. > t