Module Shared_ast.TopdefName

type t

type info = Catala_utils__Uid.Module.t list * string Catala_utils.Mark.pos

val map_info : (info -> info) -> t -> t

val compare : t -> t -> int

val equal : t -> t -> bool

val format : Stdlib.Format.formatter -> t -> unit

val to_string : t -> string

val id : t -> int

Returns the unique ID of the identifier

val original_string : t -> string

Retrieve the original id string, unaffected by renamings, etc.

val format_original : Stdlib.Format.formatter -> t -> unit

module Set : Stdlib.Set.S with type elt = t

module Map : Catala_utils.Map.S with type key = t

val fresh : ?from:t -> Catala_utils__Uid.Module.t list -> string Catala_utils.Mark.pos -> t

val path : t -> Catala_utils__Uid.Module.t list

val get_info : t -> string Catala_utils.Mark.pos

val original_info : t -> string Catala_utils.Mark.pos

val base : t -> string

Returns only the base ident name, while to_string includes the path prefix

val original_base : t -> string

val hash : strip:Catala_utils__Uid.Module.t option -> t -> Catala_utils.Hash.t

strip strips any path up to that module from the start of the path before hashing

val format_shortpath : Stdlib.Format.formatter -> t -> unit

Like format, but prints only the last path member (if any) instead of the full path

val canonical_str : Catala_utils__Uid.Module.t option -> t -> string

canonical_str current_module t returns a canonical path to t, that is, the original basename of t qualified with the last module in its path, or with the current module if implicit and that is defined.