Module Surface

module Ast : sig ... end

Abstract syntax tree built by the Catala parser

module Fill_positions : sig ... end

Fills the position tags in the AST with info about the legislative article this position belongs to.

module Lexer_common : sig ... end

Auxiliary functions used by all lexers.

module Lexer_en : sig ... end

module Lexer_fr : sig ... end

module Lexer_pl : sig ... end

module Parse_utils : sig ... end

Helpers for parsing

module Parser : sig ... end

module Parser_driver : sig ... end

Wrapping module around parser and lexer that offers the Surface.Parser_driver.parse_source_file API.

module Parser_errors : sig ... end

Interface of the module auto-generated based on "parser.messages".

module Print : sig ... end

module Tokens : sig ... end