Module Surface__Desugaring

Translation from Ast to Desugaring.Ast.

Removes syntactic sugars

Separate code from legislation

Translating expressions

val translate_op_kind : Surface.Ast.op_kind -> Dcalc.Ast.op_kind

val translate_binop : Surface.Ast.binop -> Dcalc.Ast.binop

val translate_unop : Surface.Ast.unop -> Dcalc.Ast.unop

module LiftStructFieldMap : sig ... end

module LiftEnumConstructorMap : sig ... end

val disambiguate_constructor : Surface.Name_resolution.context -> string Utils.Pos.marked list -> Utils.Pos.t -> Scopelang.Ast.EnumName.t * Scopelang.Ast.EnumConstructor.t

val translate_expr : Scopelang.Ast.ScopeName.t -> Surface.Name_resolution.context -> Surface.Ast.expression Utils.Pos.marked -> Scopelang.Ast.expr Utils.Pos.marked Bindlib.box

Usage: translate_expr scope ctxt expr

Translates expr into its desugared equivalent. scope is used to disambiguate the scope and subscopes variables than occur in the expresion

val disambiguate_match_and_build_expression : Scopelang.Ast.ScopeName.t -> Surface.Name_resolution.context -> Surface.Ast.match_cases -> Scopelang.Ast.expr Utils.Pos.marked Bindlib.box Scopelang.Ast.EnumConstructorMap.t * Scopelang.Ast.EnumName.t

Translating scope definitions

val merge_conditions : Scopelang.Ast.expr Utils.Pos.marked Bindlib.box option -> Scopelang.Ast.expr Utils.Pos.marked Bindlib.box option -> Utils.Pos.t -> Scopelang.Ast.expr Utils.Pos.marked Bindlib.box

A scope use can be annotated with a pervasive precondition, in which case this precondition has to be appended to the justifications of each definition in the subscope use. This is what this function does.

val process_default : Surface.Name_resolution.context -> Scopelang.Ast.ScopeName.t -> Desugared.Ast.ScopeDef.t Utils.Pos.marked -> Scopelang.Ast.Var.t Utils.Pos.marked option -> Scopelang.Ast.expr Utils.Pos.marked Bindlib.box option -> Desugared.Ast.RuleName.t Utils.Pos.marked option -> Surface.Ast.expression Utils.Pos.marked option -> Surface.Ast.expression Utils.Pos.marked -> Desugared.Ast.rule

Translates a surface definition into condition into a desugared Ast.rule

val process_def : Scopelang.Ast.expr Utils.Pos.marked Bindlib.box option -> Scopelang.Ast.ScopeName.t -> Surface.Name_resolution.context -> Desugared.Ast.program -> Surface.Ast.definition -> Desugared.Ast.program

Wrapper around process_default that performs some name disambiguation

val rule_to_def : Surface.Ast.rule -> Surface.Ast.definition

Translates a Ast.rule into the corresponding Ast.definition

val process_rule : Scopelang.Ast.expr Utils.Pos.marked Bindlib.box option -> Scopelang.Ast.ScopeName.t -> Surface.Name_resolution.context -> Desugared.Ast.program -> Surface.Ast.rule -> Desugared.Ast.program

Translates a Ast.rule from the surface language

val process_assert : Scopelang.Ast.expr Utils.Pos.marked Bindlib.box option -> Scopelang.Ast.ScopeName.t -> Surface.Name_resolution.context -> Desugared.Ast.program -> Surface.Ast.assertion -> Desugared.Ast.program

Translates assertions

val process_scope_use_item : Surface.Ast.expression Utils.Pos.marked option -> Scopelang.Ast.ScopeName.t -> Surface.Name_resolution.context -> Desugared.Ast.program -> Surface.Ast.scope_use_item Utils.Pos.marked -> Desugared.Ast.program

Translates a surface definition, rule or assertion

Translating top-level items

val check_unlabeled_exception : Scopelang.Ast.ScopeName.t -> Surface.Name_resolution.context -> Surface.Ast.scope_use_item Utils.Pos.marked -> unit

val process_scope_use : Surface.Name_resolution.context -> Desugared.Ast.program -> Surface.Ast.scope_use -> Desugared.Ast.program

Translates a surface scope use, which is a bunch of definitions

val desugar_program : Surface.Name_resolution.context -> Surface.Ast.program -> Desugared.Ast.program

Main function of this module