Module Surface__Lexer

Basic lexer with abbreviated syntax, and support for localised implementations.

val is_code : bool Stdlib.ref

Boolean reference, used by the lexer as the mutable state to distinguish whether it is lexing code or law.

val code_string_acc : string Stdlib.ref

Mutable string reference that accumulates the string representation of the body of code being lexed. This string representation is used in the literate programming backends to faithfully capture the spacing pattern of the original program

val update_acc : Sedlexing.lexbuf -> unit

Updates code_string_acc with the current lexeme

val raise_lexer_error : Utils.Pos.t -> string -> 'a

Error-generating helper

val token_list_language_agnostic : (string * Surface.Tokens.token) list

Associative list matching each punctuation string part of the Catala syntax with its Parser token. Same for all the input languages (English, French, etc.)

module type LocalisedLexer = sig ... end

Concise syntax with English abbreviated keywords.

include LocalisedLexer

val token_list : (string * Surface.Tokens.token) list

Same as token_list_language_agnostic, but with tokens whose string varies with the input language.

val builtins : (string * Surface.Ast.builtin_expression) list

Associative list of string to their corresponding builtins

val lex_code : Sedlexing.lexbuf -> Surface.Tokens.token

Main lexing function used in a code block

val lex_law : Sedlexing.lexbuf -> Surface.Tokens.token

Main lexing function used outside code blocks

val lexer : Sedlexing.lexbuf -> Surface.Tokens.token

Entry point of the lexer, distributes to lex_code or lex_law depending of is_code.