Module Surface__Name_resolution

Builds a context that allows for mapping each name to a precise uid, taking lexical scopes into account

module Pos = Utils.Pos

module Errors = Utils.Errors

type ident = string

type typ = Dcalc.Ast.typ

type def_context = {

var_idmap : Scopelang.Ast.Var.t Desugared.Ast.IdentMap.t;

}

Inside a definition, local variables can be introduced by functions arguments or pattern matching

type scope_context = {

var_idmap : Scopelang.Ast.ScopeVar.t Desugared.Ast.IdentMap.t;

sub_scopes_idmap : Scopelang.Ast.SubScopeName.t Desugared.Ast.IdentMap.t;

sub_scopes : Scopelang.Ast.ScopeName.t Scopelang.Ast.SubScopeMap.t;

definitions : def_context Desugared.Ast.ScopeDefMap.t;
Contains the local variables in all the definitions

}

Inside a scope, we distinguish between the variables and the subscopes.

type context = {

scope_idmap : Scopelang.Ast.ScopeName.t Desugared.Ast.IdentMap.t;

scopes : scope_context Scopelang.Ast.ScopeMap.t;

var_typs : typ Pos.marked Scopelang.Ast.ScopeVarMap.t;

}

val raise_unsupported_feature : string -> Pos.t -> 'a

val raise_unknown_identifier : string -> ident Pos.marked -> 'a

val get_var_typ : context -> Scopelang.Ast.ScopeVar.t -> typ Pos.marked

Get the type associated to an uid

val process_subscope_decl : Scopelang.Ast.ScopeName.t -> context -> Surface.Ast.scope_decl_context_scope -> context

Process a subscope declaration

val process_base_typ : Surface.Ast.base_typ Pos.marked -> Dcalc.Ast.typ Pos.marked

val process_type : Surface.Ast.typ Pos.marked -> Dcalc.Ast.typ Pos.marked

val process_data_decl : Scopelang.Ast.ScopeName.t -> context -> Surface.Ast.scope_decl_context_data -> context

Process data declaration

val process_item_decl : Scopelang.Ast.ScopeName.t -> context -> Surface.Ast.scope_decl_context_item -> context

Process an item declaration

val add_def_local_var : context -> Scopelang.Ast.ScopeName.t -> Desugared.Ast.ScopeDef.t -> ident Pos.marked -> context

Adds a binding to the context

val process_scope_decl : context -> Surface.Ast.scope_decl -> context

Process a scope declaration

val qident_to_scope_def : context -> Scopelang.Ast.ScopeName.t -> Surface.Ast.qident Pos.marked -> Desugared.Ast.ScopeDef.t

val process_scope_use : context -> Surface.Ast.scope_use -> context

val process_use_item : context -> Surface.Ast.code_item Pos.marked -> context

Process a code item : for now it only handles scope decls

val process_decl_item : context -> Surface.Ast.code_item Pos.marked -> context

Process a code item : for now it only handles scope decls

val process_code_block : context -> Surface.Ast.code_block -> (context -> Surface.Ast.code_item Pos.marked -> context) -> context

Process a code block

val process_law_article_item : context -> Surface.Ast.law_article_item -> (context -> Surface.Ast.code_item Pos.marked -> context) -> context

Process a program item

val process_law_structure : context -> Surface.Ast.law_structure -> (context -> Surface.Ast.code_item Pos.marked -> context) -> context

Process a law structure

val process_program_item : context -> Surface.Ast.program_item -> (context -> Surface.Ast.code_item Pos.marked -> context) -> context

Process a program item

val form_context : Surface.Ast.program -> context

Derive the context from metadata, in two passes

val get_var_uid : Scopelang.Ast.ScopeName.t -> context -> ident Pos.marked -> Scopelang.Ast.ScopeVar.t

Get the variable uid inside the scope given in argument

val get_subscope_uid : Scopelang.Ast.ScopeName.t -> context -> ident Pos.marked -> Scopelang.Ast.SubScopeName.t

Get the subscope uid inside the scope given in argument

val belongs_to : context -> Scopelang.Ast.ScopeVar.t -> Scopelang.Ast.ScopeName.t -> bool

Checks if the var_uid belongs to the scope scope_uid

val get_def_typ : context -> Desugared.Ast.ScopeDef.t -> typ Pos.marked