Module Surface__Parser_driver

Wrapping module around parser and lexer that offers the parse_source_file API

module I = Surface.Parser.MenhirInterpreter

Internal functions

val state : 'semantic_value I.env -> int

Returns the state number from the Menhir environment

val minimum : 'a -> 'a -> 'a -> 'a

Three-way minimum

val levenshtein_distance : string -> string -> int

Computes the levenshtein distance between two strings, used to provide error messages suggestions

val law_struct_list_to_tree : Surface.Ast.program_item list -> Surface.Ast.program_item list

After parsing, heading structure is completely flat because of the source_file_item rule. We need to tree-i-fy the flat structure, by looking at the precedence of the law headings.

val syntax_hints_style : ANSITerminal.style list

Style with which to display syntax hints in the terminal output

val raise_parser_error : Utils.Pos.t -> Utils.Pos.t option -> string -> string -> 'a

Usage: raise_parser_error error_loc last_good_loc token msg

Raises an error message featuring the error_loc position where the parser has failed, the token on which the parser has failed, and the error message msg. If available, displays last_good_loc the location of the last token correctly parsed.

val fail : Sedlexing.lexbuf -> 'semantic_value I.env -> (string * Surface.Parser.token) list -> 'semantic_value I.env option -> 'a

Usage: fail lexbuf env token_list last_input_needed

Raises an error with meaningful hints about what the parsing error was. lexbuf is the lexing buffer state at the failure point, env is the Menhir environment and last_input_needed is the last checkpoint of a valid Menhir state before the parsing error. token_list is provided by things like Lexer.token_list_language_agnostic and is used to provide suggestions of the tokens acceptable at the failure point

val loop : (unit -> Surface.Parser.token * Stdlib.Lexing.position * Stdlib.Lexing.position) -> (string * Surface.Parser.token) list -> Sedlexing.lexbuf -> Surface.Ast.source_file_or_master I.env option -> Surface.Ast.source_file_or_master I.checkpoint -> Surface.Ast.source_file_or_master

Main parsing loop

val sedlex_with_menhir : (Sedlexing.lexbuf -> Surface.Parser.token) -> (string * Surface.Parser.token) list -> (Stdlib.Lexing.position -> Surface.Ast.source_file_or_master I.checkpoint) -> Sedlexing.lexbuf -> Surface.Ast.source_file_or_master

Stub that wraps the parsing main loop and handles the Menhir/Sedlex type difference for lexbuf.

Parsing multiple files

val parse_source_file : Utils.Pos.input_file -> Utils.Cli.frontend_lang -> Surface.Ast.program

Parses a single source file

val expand_includes : string -> Surface.Ast.program_item list -> Utils.Cli.frontend_lang -> Surface.Ast.program

Expands the include directives in a parsing result, thus parsing new source files

API

val parse_top_level_file : Utils.Pos.input_file -> Utils.Cli.frontend_lang -> Surface.Ast.program