Module Utils.Astgen

module Runtime = Runtime_ocaml.Runtime

module ScopeName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module StructName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module StructFieldName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module StructMap : Stdlib.Map.S with type key = StructName.t

module EnumName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module EnumConstructor : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module EnumMap : Stdlib.Map.S with type key = EnumName.t

Abstract syntax tree for the default calculus

Abstract syntax tree

Types

type typ_lit =

| TBool

| TUnit

| TInt

| TRat

| TMoney

| TDate

| TDuration

type marked_typ = typ Marked.pos

and typ =

| TLit of typ_lit

| TTuple of marked_typ list * StructName.t option

| TEnum of marked_typ list * EnumName.t

| TArrow of marked_typ * marked_typ

| TArray of marked_typ

| TAny

Constants and operators

type date = Runtime.date

type duration = Runtime.duration

type op_kind =

| KInt

| KRat

| KMoney

| KDate

| KDuration	(*
All ops don't have a KDate and KDuration.
*)

type ternop =

| Fold

type binop =

| And

| Or

| Xor

| Add of op_kind

| Sub of op_kind

| Mult of op_kind

| Div of op_kind

| Lt of op_kind

| Lte of op_kind

| Gt of op_kind

| Gte of op_kind

| Eq

| Neq

| Map

| Concat

| Filter

type log_entry =

| VarDef of typ	(*
During code generation, we need to know the type of the variable being logged for embedding
*)

| BeginCall

| EndCall

| PosRecordIfTrueBool

type unop =

| Not

| Minus of op_kind

| Log of log_entry * Uid.MarkedString.info list

| Length

| IntToRat

| MoneyToRat

| RatToMoney

| GetDay

| GetMonth

| GetYear

| FirstDayOfMonth

| LastDayOfMonth

| RoundMoney

| RoundDecimal

type operator =

| Ternop of ternop

| Binop of binop

| Unop of unop

type except =

| ConflictError

| EmptyError

| NoValueProvided

| Crash

Generic expressions

Define a common base type for the expressions in most passes of the compiler

type desugared = [

| `Desugared

]

type scopelang = [

| `Scopelang

]

type dcalc = [

| `Dcalc

]

type lcalc = [

| `Lcalc

]

type scalc = [

| `Scalc

]

type any = [

| desugared

| scopelang

| dcalc

| lcalc

| scalc

]

type 'a glit =

| LBool : bool -> 'a glit

| LEmptyError : [< desugared | scopelang | dcalc ] glit

| LInt : Runtime.integer -> 'a glit

| LRat : Runtime.decimal -> 'a glit

| LMoney : Runtime.money -> 'a glit

| LUnit : 'a glit

| LDate : date -> 'a glit

| LDuration : duration -> 'a glit

Literals are the same throughout compilation except for the LEmptyError case which is eliminated midway through.

type ('a, 't) marked_gexpr = ( ( 'a, 't ) gexpr, 't ) Marked.t

General expressions: groups all expression cases of the different ASTs, and uses a GADT to eliminate irrelevant cases for each one. The 't annotations are also totally unconstrained at this point. The dcalc exprs, for example, are then defined with type expr = dcalc gexpr plus the annotations.

and ('a, 't) gexpr =

| ELit : 'a glit -> ( 'a, 't ) gexpr

| EApp : ( 'a, 't ) marked_gexpr * ( 'a, 't ) marked_gexpr list -> ( 'a, 't ) gexpr

| EOp : operator -> ( 'a, 't ) gexpr

| EArray : ( 'a, 't ) marked_gexpr list -> ( 'a, 't ) gexpr

| EVar : ( 'a, 't ) gexpr Bindlib.var -> ( [< desugared | scopelang | dcalc | lcalc ] as 'a, 't ) gexpr

| EAbs : ( ( 'a, 't ) gexpr, ( 'a, 't ) marked_gexpr ) Bindlib.mbinder * typ Marked.pos list -> ( [< desugared | scopelang | dcalc | lcalc ] as 'a, 't ) gexpr

| EIfThenElse : ( 'a, 't ) marked_gexpr * ( 'a, 't ) marked_gexpr * ( 'a, 't ) marked_gexpr -> ( [< desugared | scopelang | dcalc | lcalc ] as 'a, 't ) gexpr

| ETuple : ( 'a, 't ) marked_gexpr list * StructName.t option -> ( [< dcalc | lcalc ] as 'a, 't ) gexpr

| ETupleAccess : ( 'a, 't ) marked_gexpr * int * StructName.t option * typ Marked.pos list -> ( [< dcalc | lcalc ] as 'a, 't ) gexpr

| EInj : ( 'a, 't ) marked_gexpr * int * EnumName.t * typ Marked.pos list -> ( [< dcalc | lcalc ] as 'a, 't ) gexpr

| EMatch : ( 'a, 't ) marked_gexpr * ( 'a, 't ) marked_gexpr list * EnumName.t -> ( [< dcalc | lcalc ] as 'a, 't ) gexpr

| EAssert : ( 'a, 't ) marked_gexpr -> ( [< dcalc | lcalc ] as 'a, 't ) gexpr

| EDefault : ( 'a, 't ) marked_gexpr list * ( 'a, 't ) marked_gexpr * ( 'a, 't ) marked_gexpr -> ( [< desugared | scopelang | dcalc ] as 'a, 't ) gexpr

| ErrorOnEmpty : ( 'a, 't ) marked_gexpr -> ( [< desugared | scopelang | dcalc ] as 'a, 't ) gexpr

| ERaise : except -> ( lcalc as 'a, 't ) gexpr

| ECatch : ( 'a, 't ) marked_gexpr * except * ( 'a, 't ) marked_gexpr -> ( lcalc as 'a, 't ) gexpr

The expressions use the Bindlib library, based on higher-order abstract syntax

Markings

type untyped = {

pos : Pos.t;

}

type typed = {

pos : Pos.t;

ty : marked_typ;

}

type _ mark =

| Untyped : untyped -> untyped mark

| Typed : typed -> typed mark

The generic type of AST markings. Using a GADT allows functions to be polymorphic in the marking, but still do transformations on types when appropriate. Expected to fill the 't parameter of gexpr and marked_gexpr

type ('a, 'm) marked = ( 'a, 'm mark ) Marked.t

type any_marked_expr =

| AnyExpr : ( [< any ], 'm mark ) marked_gexpr -> any_marked_expr

Useful for errors and printing, for example

Higher-level program structure

Constructs scopes and programs on top of expressions. We may use the gexpr type above at some point, but at the moment this is polymorphic in the types of the expressions. Their markings are constrained to belong to the mark GADT defined above.

type scope_let_kind =

| DestructuringInputStruct	(*
let x = input.field
*)

| ScopeVarDefinition	(*
let x = error_on_empty e
*)

| SubScopeVarDefinition	(*
let s.x = fun _ -> e or let s.x = error_on_empty e for input-only subscope variables.
*)

| CallingSubScope	(*
let result = s ({ x = s.x; y = s.x; ...})
*)

| DestructuringSubScopeResults	(*
let s.x = result.x *
*)

| Assertion	(*
let _ = assert e
*)

This kind annotation signals that the let-binding respects a structural invariant. These invariants concern the shape of the expression in the let-binding, and are documented below.

type ('expr, 'm) scope_let = {

scope_let_kind : scope_let_kind;

scope_let_typ : marked_typ;

scope_let_expr : ( 'expr, 'm ) marked;

scope_let_next : ( 'expr, ( 'expr, 'm ) scope_body_expr ) Bindlib.binder;

scope_let_pos : Pos.t;

}

This type is parametrized by the expression type so it can be reused in later intermediate representations.

and ('expr, 'm) scope_body_expr =

| Result of ( 'expr, 'm ) marked

| ScopeLet of ( 'expr, 'm ) scope_let

A scope let-binding has all the information necessary to make a proper let-binding expression, plus an annotation for the kind of the let-binding that comes from the compilation of a Scopelang.Ast statement.

type ('expr, 'm) scope_body = {

scope_body_input_struct : StructName.t;

scope_body_output_struct : StructName.t;

scope_body_expr : ( 'expr, ( 'expr, 'm ) scope_body_expr ) Bindlib.binder;

}

Instead of being a single expression, we give a little more ad-hoc structure to the scope body by decomposing it in an ordered list of let-bindings, and a result expression that uses the let-binded variables. The first binder is the argument of type scope_body_input_struct.

type ('expr, 'm) scope_def = {

scope_name : ScopeName.t;

scope_body : ( 'expr, 'm ) scope_body;

scope_next : ( 'expr, ( 'expr, 'm ) scopes ) Bindlib.binder;

}

and ('expr, 'm) scopes =

| Nil

| ScopeDef of ( 'expr, 'm ) scope_def

Finally, we do the same transformation for the whole program for the kinded lets. This permit us to use bindlib variables for scopes names.

type struct_ctx = (StructFieldName.t * marked_typ) list StructMap.t

type decl_ctx = {

ctx_enums : (EnumConstructor.t * marked_typ) list EnumMap.t;

ctx_structs : struct_ctx;

}

type ('expr, 'm) program_generic = {

decl_ctx : decl_ctx;

scopes : ( 'expr, 'm ) scopes;

}