Module Utils.Astgen_utils

Functions handling the types in Astgen

Boxed constructors

val evar : ( [< Astgen.desugared | Astgen.scopelang | Astgen.dcalc | Astgen.lcalc ] as 'a, 't ) Astgen.gexpr Bindlib.var -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val etuple : ( [< Astgen.dcalc | Astgen.lcalc ] as 'a, 't ) Astgen.marked_gexpr Bindlib.box list -> Astgen.StructName.t option -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val etupleaccess : ( [< Astgen.dcalc | Astgen.lcalc ] as 'a, 't ) Astgen.marked_gexpr Bindlib.box -> int -> Astgen.StructName.t option -> Astgen.marked_typ list -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val einj : ( [< Astgen.dcalc | Astgen.lcalc ] as 'a, 't ) Astgen.marked_gexpr Bindlib.box -> int -> Astgen.EnumName.t -> Astgen.marked_typ list -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val ematch : ( [< Astgen.dcalc | Astgen.lcalc ] as 'a, 't ) Astgen.marked_gexpr Bindlib.box -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box list -> Astgen.EnumName.t -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val earray : ( 'a, 't ) Astgen.marked_gexpr Bindlib.box list -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val elit : 'a Astgen.glit -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val eabs : ( ( [< Astgen.desugared | Astgen.scopelang | Astgen.dcalc | Astgen.lcalc ] as 'a, 't ) Astgen.gexpr, ( 'a, 't ) Astgen.marked_gexpr ) Bindlib.mbinder Bindlib.box -> Astgen.marked_typ list -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val eapp : ( 'a, 't ) Astgen.marked_gexpr Bindlib.box -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box list -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val eassert : ( [< Astgen.dcalc | Astgen.lcalc ] as 'a, 't ) Astgen.marked_gexpr Bindlib.box -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val eop : Astgen.operator -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val edefault : ( [< Astgen.desugared | Astgen.scopelang | Astgen.dcalc ] as 'a, 't ) Astgen.marked_gexpr Bindlib.box list -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val eifthenelse : ( [< Astgen.desugared | Astgen.scopelang | Astgen.dcalc | Astgen.lcalc ] as 'a, 't ) Astgen.marked_gexpr Bindlib.box -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

val eerroronempty : ( [< Astgen.desugared | Astgen.scopelang | Astgen.dcalc ] as 'a, 't ) Astgen.marked_gexpr Bindlib.box -> 't -> ( 'a, 't ) Astgen.marked_gexpr Bindlib.box

----------

val map_gexpr : 'ctx -> f: ( 'ctx -> ( 'a, 't1 ) Astgen.marked_gexpr -> ( 'a, 't2 ) Astgen.marked_gexpr Bindlib.box ) -> ( ( 'a, 't1 ) Astgen.gexpr, 't2 ) Marked.t -> ( 'a, 't2 ) Astgen.marked_gexpr Bindlib.box

val map_gexpr_top_down : f: ( ( 'a, 't1 ) Astgen.marked_gexpr -> ( ( 'a, 't1 ) Astgen.gexpr, 't2 ) Marked.t ) -> ( 'a, 't1 ) Astgen.marked_gexpr -> ( 'a, 't2 ) Astgen.marked_gexpr Bindlib.box

Recursively applies f to the nodes of the expression tree. The type returned by f is hybrid since the mark at top-level has been rewritten, but not yet the marks in the subtrees.

val map_gexpr_marks : f:( 't1 -> 't2 ) -> ( 'a, 't1 ) Astgen.marked_gexpr -> ( 'a, 't2 ) Astgen.marked_gexpr Bindlib.box

val fold_left_scope_lets : f:( 'a -> ( 'expr, 'm ) Astgen.scope_let -> 'expr Bindlib.var -> 'a ) -> init:'a -> ( 'expr, 'm ) Astgen.scope_body_expr -> 'a

Usage: fold_left_scope_lets ~f:(fun acc scope_let scope_let_var -> ...) ~init scope_lets, where scope_let_var is the variable bound to the scope let in the next scope lets to be examined.

val fold_right_scope_lets : f:( ( 'expr1, 'm1 ) Astgen.scope_let -> 'expr1 Bindlib.var -> 'a -> 'a ) -> init:( ( 'expr1, 'm1 ) Astgen.marked -> 'a ) -> ( 'expr1, 'm1 ) Astgen.scope_body_expr -> 'a

Usage: fold_right_scope_lets ~f:(fun scope_let scope_let_var acc -> ...) ~init scope_lets, where scope_let_var is the variable bound to the scope let in the next scope lets to be examined (which are before in the program order).

val map_exprs_in_scope_lets : f: ( ( 'expr1, 'm1 ) Astgen.marked -> ( 'expr2, 'm2 ) Astgen.marked Bindlib.box ) -> varf:( 'expr1 Bindlib.var -> 'expr2 Bindlib.var ) -> ( 'expr1, 'm1 ) Astgen.scope_body_expr -> ( 'expr2, 'm2 ) Astgen.scope_body_expr Bindlib.box

val fold_left_scope_defs : f:( 'a -> ( 'expr1, 'm1 ) Astgen.scope_def -> 'expr1 Bindlib.var -> 'a ) -> init:'a -> ( 'expr1, 'm1 ) Astgen.scopes -> 'a

Usage: fold_left_scope_defs ~f:(fun acc scope_def scope_var -> ...) ~init scope_def, where scope_var is the variable bound to the scope in the next scopes to be examined.

val fold_right_scope_defs : f:( ( 'expr1, 'm1 ) Astgen.scope_def -> 'expr1 Bindlib.var -> 'a -> 'a ) -> init:'a -> ( 'expr1, 'm1 ) Astgen.scopes -> 'a

Usage: fold_right_scope_defs ~f:(fun scope_def scope_var acc -> ...) ~init scope_def, where scope_var is the variable bound to the scope in the next scopes to be examined (which are before in the program order).

val map_scope_defs : f: ( ( 'expr, 'm ) Astgen.scope_def -> ( 'expr, 'm ) Astgen.scope_def Bindlib.box ) -> ( 'expr, 'm ) Astgen.scopes -> ( 'expr, 'm ) Astgen.scopes Bindlib.box

val map_exprs_in_scopes : f: ( ( 'expr1, 'm1 ) Astgen.marked -> ( 'expr2, 'm2 ) Astgen.marked Bindlib.box ) -> varf:( 'expr1 Bindlib.var -> 'expr2 Bindlib.var ) -> ( 'expr1, 'm1 ) Astgen.scopes -> ( 'expr2, 'm2 ) Astgen.scopes Bindlib.box

This is the main map visitor for all the expressions inside all the scopes of the program.