Module Utils.String_common

Helper functions used for string manipulation.

val to_ascii : string -> string

Removes all non-ASCII diacritics from a string by converting them to their base letter in the Latin alphabet.

val is_uppercase_ascii : char -> bool

is_uppercase c returns if c is in the set 'A'...'Z'.

val begins_with_uppercase : string -> bool

begins_with_uppercase s returns if the first letter of s is_uppercase_ascii. If s is empty returns false.

val to_snake_case : string -> string

Converts CamlCase into snake_case after removing Remove all diacritics on Latin letters.

val to_camel_case : string -> string

Converts snake_case into CamlCase after removing Remove all diacritics on Latin letters.