Module Utils__Cli

type backend_lang =

| En

| Fr

| Pl

Configuration globals

val source_files : string list Stdlib.ref

Source files to be compiled

val locale_lang : backend_lang Stdlib.ref

val contents : string Stdlib.ref

val debug_flag : bool Stdlib.ref

val style_flag : bool Stdlib.ref

Styles the terminal output

val optimize_flag : bool Stdlib.ref

val max_prec_digits : int Stdlib.ref

Max number of digits to show for decimal results

val trace_flag : bool Stdlib.ref

val disable_counterexamples : bool Stdlib.ref

Disables model-generated counterexamples for proofs that fail.

val avoid_exceptions_flag : bool Stdlib.ref

Avoids using try ... with exceptions when compiling the default calculus.

CLI terms

val file : string Cmdliner.Term.t

val debug : bool Cmdliner.Term.t

val unstyled : bool Cmdliner.Term.t

val trace_opt : bool Cmdliner.Term.t

val wrap_weaved_output : bool Cmdliner.Term.t

val backend : string Cmdliner.Term.t

type backend_option =

| Dcalc

| Html

| Interpret

| Latex

| Lcalc

| Makefile

| OCaml

| Proof

| Python

| Scalc

| Scopelang

| Typecheck

val language : string option Cmdliner.Term.t

val max_prec_digits_opt : int option Cmdliner.Term.t

val ex_scope : string option Cmdliner.Term.t

val output : string option Cmdliner.Term.t

val catala_t : (string -> bool -> bool -> bool -> bool -> string -> string option -> int option -> bool -> bool -> bool -> string option -> string option -> 'a) -> 'a Cmdliner.Term.t

Main entry point: catala_t file debug unstyled wrap_weaved_output avoid_exceptions backend language max_prec_digits_opt trace_opt disable_counterexamples optimize ex_scope output

val version : string

val info : Cmdliner.Term.info

Terminal formatting

Markers

val with_style : ANSITerminal.style list -> ('a, unit, string) Stdlib.format -> 'a

val format_with_style : ANSITerminal.style list -> Stdlib.Format.formatter -> string -> unit

val debug_marker : unit -> string

val error_marker : unit -> string

val warning_marker : unit -> string

val result_marker : unit -> string

val log_marker : unit -> string

Printers

val concat_with_line_depending_prefix_and_suffix : (int -> string) -> (int -> string) -> string list -> string

val add_prefix_to_each_line : string -> (int -> string) -> string

The int argument of the prefix corresponds to the line number, starting at 0

val debug_print : ('a, Stdlib.out_channel, unit) Stdlib.format -> 'a

val debug_format : ('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a

val error_print : ('a, Stdlib.out_channel, unit) Stdlib.format -> 'a

val warning_print : ('a, Stdlib.out_channel, unit) Stdlib.format -> 'a

val result_print : ('a, Stdlib.out_channel, unit) Stdlib.format -> 'a

val result_format : ('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a

val log_print : ('a, Stdlib.out_channel, unit) Stdlib.format -> 'a

val log_format : ('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a