Module Verification__Conditions

Generates verification conditions from scope definitions

type verification_condition_kind =

| NoEmptyError
This verification condition checks whether a definition never returns an empty error

| NoOverlappingExceptions
This verification condition checks whether a definition never returns a conflict error

type verification_condition = {

vc_guard : Dcalc.Ast.expr Utils.Pos.marked;
This expression should have type bool

vc_kind : verification_condition_kind;

vc_scope : Dcalc.Ast.ScopeName.t;

vc_variable : Dcalc.Ast.Var.t Utils.Pos.marked;

vc_free_vars_typ : Dcalc.Ast.typ Utils.Pos.marked Dcalc.Ast.VarMap.t;
Types of the locally free variables in vc_guard. The types of other free variables linked to scope variables can be obtained with Dcalc.Ast.variable_types.

}

val generate_verification_conditions : Dcalc.Ast.program -> verification_condition list