Module Catala_utils.Cli

type backend_lang =

| En

| Fr

| Pl

type backend_option = [

| `Latex

| `Makefile

| `Html

| `Interpret

| `Interpret_Lcalc

| `Typecheck

| `OCaml

| `Python

| `Scalc

| `Lcalc

| `Dcalc

| `Scopelang

| `Exceptions

| `Proof

]

type when_enum =

| Auto

| Always

| Never

The usual auto/always/never option argument

val languages : (string * backend_lang) list

val language_code : backend_lang -> string

Returns the lowercase two-letter language code

Configuration globals

val source_files : string list Stdlib.ref

Source files to be compiled

val locale_lang : backend_lang Stdlib.ref

val contents : string Stdlib.ref

val debug_flag : bool Stdlib.ref

val style_flag : when_enum Stdlib.ref

Styles the terminal output

val optimize_flag : bool Stdlib.ref

val max_prec_digits : int Stdlib.ref

Max number of digits to show for decimal results

val trace_flag : bool Stdlib.ref

val disable_warnings_flag : bool Stdlib.ref

val check_invariants_flag : bool Stdlib.ref

Check structural invariants on the AST.

val disable_counterexamples : bool Stdlib.ref

Disables model-generated counterexamples for proofs that fail.

val avoid_exceptions_flag : bool Stdlib.ref

Avoids using try ... with exceptions when compiling the default calculus.

type message_format_enum =

| Human

| GNU	(*
Format of error and warning messages output by the compiler.
*)

val message_format_flag : message_format_enum Stdlib.ref

CLI terms

val file : string Cmdliner.Term.t

val debug : bool Cmdliner.Term.t

val unstyled : bool Cmdliner.Term.t

val trace_opt : bool Cmdliner.Term.t

val check_invariants_opt : bool Cmdliner.Term.t

val wrap_weaved_output : bool Cmdliner.Term.t

val print_only_law : bool Cmdliner.Term.t

val plugins_dirs : string list Cmdliner.Term.t

val language : string option Cmdliner.Term.t

val max_prec_digits_opt : int option Cmdliner.Term.t

val ex_scope : string option Cmdliner.Term.t

val output : string option Cmdliner.Term.t

type global_options = {

debug : bool;

color : when_enum;

message_format : message_format_enum;

wrap_weaved_output : bool;

avoid_exceptions : bool;

plugins_dirs : string list;

language : string option;

max_prec_digits : int option;

trace : bool;

disable_warnings : bool;

disable_counterexamples : bool;

check_invariants : bool;

optimize : bool;

ex_scope : string option;

ex_variable : string option;

output_file : string option;

closure_conversion : bool;

print_only_law : bool;

link_modules : string list;

}

Command-line application

val global_options : global_options Cmdliner.Term.t

val catala_t : ?extra:int Cmdliner.Cmd.t list -> ( backend_option -> string -> global_options -> int ) -> int Cmdliner.Cmd.t

Main entry point: catala_t file options

val set_option_globals : global_options -> unit

val version : string

val info : Cmdliner.Cmd.info