Module Flags.Global

val flags : options Cmdliner.Term.t

Global flags available to all commands. Note that parsing this term also performs some side-effects into GlobalRefs and sets up signal/error processing. Sets input_file to FileName "-", use options for the full parser

val options : options Cmdliner.Term.t

flags plus an additional positional argument for the input file