Module Catala_utils.Message

Interface for emitting compiler messages.

All messages are expected to use the Format module. Flush, "@?", "@.", "%!" etc. are not supposed to be used outside of this module.

WARNING: this module performs side-effects at load time, adding support for ocolor tags (e.g. "@{<blue>text@}") to the standard string formatter used by e.g. Format.sprintf. (In this case, the tags are ignored, for color output you should use the functions of this module that toggle support depending on cli flags and terminal support).

Message content

type content_type =

| Error

| Warning

| Debug

| Log

| Result

module Content : sig ... end

This functions emits the message according to the emission type defined by Cli.message_format_flag.

Error exception

exception CompilerError of Content.t

Common error raising

val raise_spanned_error : ?span_msg:Content.message -> ?suggestion:string list -> Pos.t -> ( 'a, Stdlib.Format.formatter, unit, 'b ) Stdlib.format4 -> 'a

val raise_multispanned_error_full : ?suggestion:string list -> (Content.message option * Pos.t) list -> ( 'a, Stdlib.Format.formatter, unit, 'b ) Stdlib.format4 -> 'a

val raise_multispanned_error : ?suggestion:string list -> (string option * Pos.t) list -> ( 'a, Stdlib.Format.formatter, unit, 'b ) Stdlib.format4 -> 'a

val raise_error : ( 'a, Stdlib.Format.formatter, unit, 'b ) Stdlib.format4 -> 'a

val raise_internal_error : ( 'a, Stdlib.Format.formatter, unit, 'b ) Stdlib.format4 -> 'a

val assert_internal_error : bool -> ( 'a, Stdlib.Format.formatter, unit, unit, unit, unit ) Stdlib.format6 -> 'a

Common warning emission

val emit_multispanned_warning : (Content.message option * Pos.t) list -> ( 'a, Stdlib.Format.formatter, unit ) Stdlib.format -> 'a

val emit_spanned_warning : ?span_msg:Content.message -> Pos.t -> ( 'a, Stdlib.Format.formatter, unit ) Stdlib.format -> 'a

val emit_warning : ( 'a, Stdlib.Format.formatter, unit ) Stdlib.format -> 'a

Common log emission

val emit_log : ( 'a, Stdlib.Format.formatter, unit ) Stdlib.format -> 'a

Common debug emission

val emit_debug : ( 'a, Stdlib.Format.formatter, unit ) Stdlib.format -> 'a

Common result emission

val emit_result : ( 'a, Stdlib.Format.formatter, unit ) Stdlib.format -> 'a

Some formatting helpers

val unformat : ( Stdlib.Format.formatter -> unit ) -> string

Converts f to a string, discarding formatting and skipping newlines and indents

val formatter_of_out_channel : Stdlib.out_channel -> Stdlib.Format.formatter

Creates a new formatter from the given out channel, with correct handling of the ocolor tags. Actual use of escape codes in the output depends on Cli.style_flag -- and wether the channel is a tty if that is set to auto.