Module Catala_utils.Pos

Source code position

type t

A position in the source code is a file, as well as begin and end location of the form col:line

Constructor and getters

val from_lpos : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> t

val from_info : string -> int -> int -> int -> int -> t

val overwrite_law_info : t -> string list -> t

val get_law_info : t -> string list

val get_start_line : t -> int

val get_start_column : t -> int

val get_end_line : t -> int

val get_end_column : t -> int

val get_file : t -> string

val join : t -> t -> t

Returns the smallest range including both supplied ranges.

raises Invalid_argument
if they don't belong to the same file. The law position used is the one of the earliest position.

type input_file =

| FileName of string

| Contents of string

Formatters

val to_string : t -> string

Formats a position like this:

in file <file>, from <start_line>:<start_col> to <end_line>:<end_col>

val to_string_short : t -> string

Formats a position like this:

<file>;<start_line>:<start_col>--<end_line>:<end_col>

val retrieve_loc_text : t -> string

Open the file corresponding to the position and retrieves the text concerned by the position

val no_pos : t

Placeholder position