Module Catala_utils.Uid

Global identifiers factories using a generative functor

module type Info = sig ... end

The information carried in global identifiers

module MarkedString : Info with type info = string Marked.pos

The only kind of information carried in Catala identifiers is the original string of the identifier annotated with the position where it is declared or used.

module type Id = sig ... end

Identifiers have abstract types, but are comparable so they can be used as keys in maps or sets. Their underlying information can be retrieved at any time.

module Make (X : Info) () : Id with type info = X.info

This is the generative functor that ensures that two modules resulting from two different calls to Make will be viewed as different types t by the OCaml typechecker. Prevents mixing up different sorts of identifiers.

module Gen () : Id with type info = MarkedString.info

Shortcut for creating a kind of uids over marked strings