Module Uid.MarkedString

The only kind of information carried in Catala identifiers is the original string of the identifier annotated with the position where it is declared or used.

type info = string Marked.pos

val to_string : info -> string

val format : Stdlib.Format.formatter -> info -> unit

val equal : info -> info -> bool

Equality disregards position

val compare : info -> info -> int

Comparison disregards position