Module Dcalc

module Ast : sig ... end

Abstract syntax tree of the default calculus intermediate representation

module From_scopelang : sig ... end

Scope language to default calculus translator

module Invariants : sig ... end

This file makes explicit few structural invariants of the dcalc asbtract syntax tree. Those invariants have been checked on all tests and examples of catala. The behavior of the compiler on programs that don't follow those invariant in undefined.