Module Dcalc.Ast

Abstract syntax tree of the default calculus intermediate representation

module ScopeName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module StructName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module StructFieldName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module StructMap : Stdlib.Map.S with type key = StructName.t

module EnumName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module EnumConstructor : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module EnumMap : Stdlib.Map.S with type key = EnumName.t

Abstract syntax tree for the default calculus

Abstract syntax tree

type typ_lit =

| TBool

| TUnit

| TInt

| TRat

| TMoney

| TDate

| TDuration

type typ =

| TLit of typ_lit

| TTuple of typ Utils.Pos.marked list * StructName.t option

| TEnum of typ Utils.Pos.marked list * EnumName.t

| TArrow of typ Utils.Pos.marked * typ Utils.Pos.marked

| TArray of typ Utils.Pos.marked

| TAny

type date = Runtime.date

type duration = Runtime.duration

type lit =

| LBool of bool

| LEmptyError

| LInt of Runtime.integer

| LRat of Runtime.decimal

| LMoney of Runtime.money

| LUnit

| LDate of date

| LDuration of duration

type op_kind =

| KInt

| KRat

| KMoney

| KDate

| KDuration	(*
All ops don't have a KDate and KDuration.
*)

type ternop =

| Fold

type binop =

| And

| Or

| Xor

| Add of op_kind

| Sub of op_kind

| Mult of op_kind

| Div of op_kind

| Lt of op_kind

| Lte of op_kind

| Gt of op_kind

| Gte of op_kind

| Eq

| Neq

| Map

| Concat

| Filter

type log_entry =

| VarDef of typ	(*
During code generation, we need to know the type of the variable being logged for embedding
*)

| BeginCall

| EndCall

| PosRecordIfTrueBool

type unop =

| Not

| Minus of op_kind

| Log of log_entry * Utils.Uid.MarkedString.info list

| Length

| IntToRat

| GetDay

| GetMonth

| GetYear

| RoundMoney

type operator =

| Ternop of ternop

| Binop of binop

| Unop of unop

type expr =

| EVar of expr Bindlib.var Utils.Pos.marked

| ETuple of expr Utils.Pos.marked list * StructName.t option	(*
The MarkedString.info is the former struct field name
*)

| ETupleAccess of expr Utils.Pos.marked * int * StructName.t option * typ Utils.Pos.marked list	(*
The MarkedString.info is the former struct field name
*)

| EInj of expr Utils.Pos.marked * int * EnumName.t * typ Utils.Pos.marked list	(*
The MarkedString.info is the former enum case name
*)

| EMatch of expr Utils.Pos.marked * expr Utils.Pos.marked list * EnumName.t	(*
The MarkedString.info is the former enum case name
*)

| EArray of expr Utils.Pos.marked list

| ELit of lit

| EAbs of ( expr, expr Utils.Pos.marked ) Bindlib.mbinder Utils.Pos.marked * typ Utils.Pos.marked list

| EApp of expr Utils.Pos.marked * expr Utils.Pos.marked list

| EAssert of expr Utils.Pos.marked

| EOp of operator

| EDefault of expr Utils.Pos.marked list * expr Utils.Pos.marked * expr Utils.Pos.marked

| EIfThenElse of expr Utils.Pos.marked * expr Utils.Pos.marked * expr Utils.Pos.marked

| ErrorOnEmpty of expr Utils.Pos.marked

The expressions use the Bindlib library, based on higher-order abstract syntax

type struct_ctx = (StructFieldName.t * typ Utils.Pos.marked) list StructMap.t

type enum_ctx = (EnumConstructor.t * typ Utils.Pos.marked) list EnumMap.t

type decl_ctx = {

ctx_enums : enum_ctx;

ctx_structs : struct_ctx;

}

type binder = ( expr, expr Utils.Pos.marked ) Bindlib.binder

type scope_let_kind =

| DestructuringInputStruct	(*
let x = input.field
*)

| ScopeVarDefinition	(*
let x = error_on_empty e
*)

| SubScopeVarDefinition	(*
let s.x = fun _ -> e or let s.x = error_on_empty e for input-only subscope variables.
*)

| CallingSubScope	(*
let result = s ({ x = s.x; y = s.x; ...})
*)

| DestructuringSubScopeResults	(*
let s.x = result.x *
*)

| Assertion	(*
let _ = assert e
*)

This kind annotation signals that the let-binding respects a structural invariant. These invariants concern the shape of the expression in the let-binding, and are documented below.

type 'expr scope_let = {

scope_let_kind : scope_let_kind;

scope_let_typ : typ Utils.Pos.marked;

scope_let_expr : 'expr Utils.Pos.marked;

scope_let_next : ( 'expr, 'expr scope_body_expr ) Bindlib.binder;

scope_let_pos : Utils.Pos.t;

}

This type is parametrized by the expression type so it can be reused in later intermediate representations.

and 'expr scope_body_expr =

| Result of 'expr Utils.Pos.marked

| ScopeLet of 'expr scope_let

A scope let-binding has all the information necessary to make a proper let-binding expression, plus an annotation for the kind of the let-binding that comes from the compilation of a Scopelang.Ast statement.

type 'expr scope_body = {

scope_body_input_struct : StructName.t;

scope_body_output_struct : StructName.t;

scope_body_expr : ( 'expr, 'expr scope_body_expr ) Bindlib.binder;

}

Instead of being a single expression, we give a little more ad-hoc structure to the scope body by decomposing it in an ordered list of let-bindings, and a result expression that uses the let-binded variables. The first binder is the argument of type scope_body_input_struct.

type 'expr scope_def = {

scope_name : ScopeName.t;

scope_body : 'expr scope_body;

scope_next : ( 'expr, 'expr scopes ) Bindlib.binder;

}

and 'a scopes =

| Nil

| ScopeDef of 'a scope_def

Finally, we do the same transformation for the whole program for the kinded lets. This permit us to use bindlib variables for scopes names.

type program = {

decl_ctx : decl_ctx;

scopes : expr scopes;

}

Helpers

Boxed constructors

val evar : expr Bindlib.var -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val etuple : expr Utils.Pos.marked Bindlib.box list -> StructName.t option -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val etupleaccess : expr Utils.Pos.marked Bindlib.box -> int -> StructName.t option -> typ Utils.Pos.marked list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val einj : expr Utils.Pos.marked Bindlib.box -> int -> EnumName.t -> typ Utils.Pos.marked list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val ematch : expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box list -> EnumName.t -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val earray : expr Utils.Pos.marked Bindlib.box list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val elit : lit -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val eabs : ( expr, expr Utils.Pos.marked ) Bindlib.mbinder Bindlib.box -> Utils.Pos.t -> typ Utils.Pos.marked list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val eapp : expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val eassert : expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val eop : operator -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val edefault : expr Utils.Pos.marked Bindlib.box list -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val eifthenelse : expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val eerroronempty : expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

Program traversal

Be careful when using these traversal functions, as the bound variables they open will be different at each traversal.

val map_expr : 'a -> f:( 'a -> expr Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box ) -> expr Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box

If you want to apply a map transform to an expression, you can save up writing a painful match over all the cases of the AST. For instance, if you want to remove all errors on empty, you can write

let remove_error_empty =
  let rec f () e =
    match Pos.unmark e with
    | ErrorOnEmpty e1 -> map_expr () f e1
    | _ -> map_expr () f e
  in
  f () e

The first argument of map_expr is an optional context that you can carry around during your map traversal.

val fold_scope_lets : f:( 'a -> 'expr scope_let -> 'a ) -> init:'a -> 'expr scope_body_expr -> 'a

val fold_scope_defs : f:( 'a -> 'expr scope_def -> 'a ) -> init:'a -> 'expr scopes -> 'a

Variables

module Var : sig ... end

module VarMap : Stdlib.Map.S with type key = Var.t

module VarSet : Stdlib.Set.S with type elt = Var.t

val free_vars_expr : expr Utils.Pos.marked -> VarSet.t

val free_vars_scope_body_expr : expr scope_body_expr -> VarSet.t

val free_vars_scope_body : expr scope_body -> VarSet.t

val free_vars_scopes : expr scopes -> VarSet.t

type vars = expr Bindlib.mvar

Boxed term constructors

val make_var : Var.t Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box

val make_abs : vars -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> typ Utils.Pos.marked list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_app : expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_let_in : Var.t -> typ Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val box_expr : expr Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box

Other

val empty_thunked_term : expr Utils.Pos.marked

val is_value : expr Utils.Pos.marked -> bool

val equal_exprs : expr Utils.Pos.marked -> expr Utils.Pos.marked -> bool

Determines if two expressions are equal, omitting their position information

AST manipulation helpers

type 'expr make_let_in_sig = 'expr Bindlib.var -> typ Utils.Pos.marked -> 'expr Utils.Pos.marked Bindlib.box -> 'expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> 'expr Utils.Pos.marked Bindlib.box

type 'expr make_abs_sig = 'expr Bindlib.mvar -> 'expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> typ Utils.Pos.marked list -> Utils.Pos.t -> 'expr Utils.Pos.marked Bindlib.box

type 'expr box_expr_sig = 'expr Utils.Pos.marked -> 'expr Utils.Pos.marked Bindlib.box

val build_whole_scope_expr : box_expr:'expr box_expr_sig -> make_abs:'expr make_abs_sig -> make_let_in:'expr make_let_in_sig -> decl_ctx -> 'expr scope_body -> Utils.Pos.t -> 'expr Utils.Pos.marked Bindlib.box

Usage: build_whole_scope_expr ctx body scope_position where scope_position corresponds to the line of the scope declaration for instance.

type 'expr scope_name_or_var =

| ScopeName of ScopeName.t

| ScopeVar of 'expr Bindlib.var

val unfold_scopes : box_expr:'expr box_expr_sig -> make_abs:'expr make_abs_sig -> make_let_in:'expr make_let_in_sig -> decl_ctx -> 'expr scopes -> 'expr scope_name_or_var -> 'expr Utils.Pos.marked Bindlib.box

val build_whole_program_expr : program -> ScopeName.t -> expr Utils.Pos.marked Bindlib.box

Usage: build_whole_program_expr program main_scope builds an expression corresponding to the main program and returning the main scope as a function.

val expr_size : expr Utils.Pos.marked -> int

Used by the optimizer to know when to stop

val remove_logging_calls : expr Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box

Removes all calls to Log unary operators in the AST, replacing them by their argument.