Module Dcalc__Interpreter

Reference interpreter for the default calculus

val empty_thunked_term : Dcalc.Ast.expr Utils.Pos.marked

val interpret_program : Dcalc.Ast.decl_ctx -> Dcalc.Ast.expr Utils.Pos.marked -> (Utils.Uid.MarkedString.info * Dcalc.Ast.expr Utils.Pos.marked) list

Interpret a program. This function expects an expression typed as a function whose argument are all thunked. The function is executed by providing for each argument a thunked empty default.