Module Desugared.Ast

Abstract syntax tree of the desugared representation

Names, Maps and Keys

module IdentMap : Stdlib.Map.S with type key = Stdlib.String.t

module RuleName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module RuleMap : Stdlib.Map.S with type key = RuleName.t

module RuleSet : Stdlib.Set.S with type elt = RuleName.t

module LabelName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module LabelMap : Stdlib.Map.S with type key = LabelName.t

module LabelSet : Stdlib.Set.S with type elt = LabelName.t

module StateName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module ScopeVar : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module ScopeVarSet : Stdlib.Set.S with type elt = ScopeVar.t

module ScopeVarMap : Stdlib.Map.S with type key = ScopeVar.t

module ScopeDef : sig ... end

Inside a scope, a definition can refer either to a scope def, or a subscope def

module ScopeDefMap : Stdlib.Map.S with type key = ScopeDef.t

module ScopeDefSet : Stdlib.Set.S with type elt = ScopeDef.t

AST

type location =

| ScopeVar of ScopeVar.t Utils.Marked.pos * StateName.t option

| SubScopeVar of Scopelang.Ast.ScopeName.t * Scopelang.Ast.SubScopeName.t Utils.Marked.pos * ScopeVar.t Utils.Marked.pos

Expressions

module LocationSet : Stdlib.Set.S with type elt = location Utils.Marked.pos

type marked_expr = expr Utils.Marked.pos

The expressions use the Bindlib library, based on higher-order abstract syntax

and expr =

| ELocation of location

| EVar of expr Bindlib.var

| EStruct of Scopelang.Ast.StructName.t * marked_expr Scopelang.Ast.StructFieldMap.t

| EStructAccess of marked_expr * Scopelang.Ast.StructFieldName.t * Scopelang.Ast.StructName.t

| EEnumInj of marked_expr * Scopelang.Ast.EnumConstructor.t * Scopelang.Ast.EnumName.t

| EMatch of marked_expr * Scopelang.Ast.EnumName.t * marked_expr Scopelang.Ast.EnumConstructorMap.t

| ELit of Dcalc.Ast.lit

| EAbs of ( expr, marked_expr ) Bindlib.mbinder * Scopelang.Ast.typ Utils.Marked.pos list

| EApp of marked_expr * marked_expr list

| EOp of Dcalc.Ast.operator

| EDefault of marked_expr list * marked_expr * marked_expr

| EIfThenElse of marked_expr * marked_expr * marked_expr

| EArray of marked_expr list

| ErrorOnEmpty of marked_expr

module ExprMap : Stdlib.Map.S with type key = expr

Variable helpers

module Var : sig ... end

module VarMap : Stdlib.Map.S with type key = Var.t

type vars = expr Bindlib.mvar

val make_var : Var.t Utils.Marked.pos -> expr Utils.Marked.pos Bindlib.box

val make_abs : vars -> expr Utils.Marked.pos Bindlib.box -> Scopelang.Ast.typ Utils.Marked.pos list -> Utils.Pos.t -> expr Utils.Marked.pos Bindlib.box

val make_app : expr Utils.Marked.pos Bindlib.box -> expr Utils.Marked.pos Bindlib.box list -> Utils.Pos.t -> expr Utils.Marked.pos Bindlib.box

val make_let_in : Var.t -> Scopelang.Ast.typ Utils.Marked.pos -> expr Utils.Marked.pos Bindlib.box -> expr Utils.Marked.pos Bindlib.box -> expr Utils.Marked.pos Bindlib.box

Rules and scopes

type rule = {

rule_id : RuleName.t;

rule_just : expr Utils.Marked.pos Bindlib.box;

rule_cons : expr Utils.Marked.pos Bindlib.box;

rule_parameter : (Var.t * Scopelang.Ast.typ Utils.Marked.pos) option;

rule_exception_to_rules : RuleSet.t Utils.Marked.pos;

}

module Rule : Stdlib.Set.OrderedType with type t = rule

val empty_rule : Utils.Pos.t -> Scopelang.Ast.typ Utils.Marked.pos option -> rule

val always_false_rule : Utils.Pos.t -> Scopelang.Ast.typ Utils.Marked.pos option -> rule

type assertion = expr Utils.Marked.pos Bindlib.box

type variation_typ =

| Increasing

| Decreasing

type reference_typ =

| Decree

| Law

type meta_assertion =

| FixedBy of reference_typ Utils.Marked.pos

| VariesWith of unit * variation_typ Utils.Marked.pos option

type scope_def = {

scope_def_rules : rule RuleMap.t;

scope_def_typ : Scopelang.Ast.typ Utils.Marked.pos;

scope_def_is_condition : bool;

scope_def_io : Scopelang.Ast.io;

scope_def_label_groups : RuleSet.t LabelMap.t;

}

type var_or_states =

| WholeVar

| States of StateName.t list

type scope = {

scope_vars : var_or_states ScopeVarMap.t;

scope_sub_scopes : Scopelang.Ast.ScopeName.t Scopelang.Ast.SubScopeMap.t;

scope_uid : Scopelang.Ast.ScopeName.t;

scope_defs : scope_def ScopeDefMap.t;

scope_assertions : assertion list;

scope_meta_assertions : meta_assertion list;

}

type program = {

program_scopes : scope Scopelang.Ast.ScopeMap.t;

program_enums : Scopelang.Ast.enum_ctx;

program_structs : Scopelang.Ast.struct_ctx;

}

Helpers

val locations_used : expr Utils.Marked.pos -> LocationSet.t

val free_variables : rule RuleMap.t -> Utils.Pos.t ScopeDefMap.t