Module Lcalc.Closure_conversion

TODO: This version is not yet debugged and ought to be specialized when Lcalc has more structure.

type closure = {

name : Ast.Var.t;

expr : Ast.expr Utils.Pos.marked Bindlib.box;

}

type ctx = {

name_context : string;

globally_bound_vars : Lcalc.Ast.VarSet.t;

}

val closure_conversion_expr : ctx -> Ast.expr Utils.Pos.marked -> Ast.expr Utils.Pos.marked Bindlib.box * Lcalc.Ast.VarSet.t

Returns the expression with closed closures and the set of free variables inside this new expression. Implementation guided by http://gallium.inria.fr/~fpottier/mpri/cours04.pdf#page=9.

val closure_conversion : Ast.program -> Ast.program Bindlib.box * closure list