Module Literate.Html

This modules weaves the source code and the legislative text together into a document that law professionals can understand.

Helpers

val wrap_html : string list -> Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> ( Stdlib.Format.formatter -> unit ) -> unit

Usage: wrap_html source_files language fmt wrapped

Prints an HTML complete page structure around the wrapped content.

API

val ast_to_html : Catala_utils.Cli.backend_lang -> print_only_law:bool -> Stdlib.Format.formatter -> Surface.Ast.program -> unit