Module Literate__Html

This modules weaves the source code and the legislative text together into a document that law professionals can understand.

module Pos = Utils.Pos

module Cli = Utils.Cli

module Errors = Utils.Errors

module A = Surface.Ast

module P = Stdlib.Printf

module R = Re.Pcre

module C = Cli

val pre_html : string -> string

val raise_failed_pygments : string -> int -> 'a

val wrap_html : string list -> string option -> Cli.backend_lang -> Stdlib.Format.formatter -> (Stdlib.Format.formatter -> unit) -> unit

val pygmentize_code : string Pos.marked -> C.backend_lang -> string option -> string

type program_state =

| InsideArticle

| OutsideArticle

val law_article_item_to_html : string option -> C.backend_lang -> Stdlib.Format.formatter -> A.law_article_item -> unit

val law_structure_to_html : string option -> C.backend_lang -> Stdlib.Format.formatter -> A.law_structure -> unit

val program_item_to_html : string option -> C.backend_lang -> Stdlib.Format.formatter -> A.program_item -> unit

val ast_to_html : string option -> C.backend_lang -> Stdlib.Format.formatter -> A.program -> unit