Module Literate__Latex

This modules weaves the source code and the legislative text together into a document that law professionals can understand.

Helpers

val wrap_latex : string list -> Utils.Cli.backend_lang -> Stdlib.Format.formatter -> (Stdlib.Format.formatter -> unit) -> unit

Usage: wrap_latex source_files language fmt wrapped

Prints an LaTeX complete documùent structure around the wrapped content.

API

val ast_to_latex : Utils.Cli.backend_lang -> Stdlib.Format.formatter -> Surface.Ast.program -> unit