Module Python

This file is only for demonstration purposes, showing a trivial use of backend plugins for Catala.

The code for the Python backend already has first-class support, so there would be no reason to use this plugin instead

val run : string list -> string option -> bool -> bool -> bool -> bool -> Catala_utils.Cli.options -> unit

val term : ( Catala_utils.Cli.options -> unit ) Cmdliner.Term.t