Module Scopelang

module Ast : sig ... end

Abstract syntax tree of the scope language

module Dependency : sig ... end

Graph representation of the dependencies between scopes in the Catala program. Vertices are functions, x -> y if x is used in the definition of y.

module From_desugared : sig ... end

Translation from Desugared.Ast to Scopelang.Ast

module Print : sig ... end