Module Scopelang.Ast

Abstract syntax tree of the scope language

Identifiers

module ScopeName = Dcalc.Ast.ScopeName

module ScopeNameSet : Stdlib.Set.S with type elt = ScopeName.t

module ScopeMap : Stdlib.Map.S with type key = ScopeName.t

module SubScopeName : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module SubScopeNameSet : Stdlib.Set.S with type elt = SubScopeName.t

module SubScopeMap : Stdlib.Map.S with type key = SubScopeName.t

module ScopeVar : Utils.Uid.Id with type info = Utils.Uid.MarkedString.info

module ScopeVarSet : Stdlib.Set.S with type elt = ScopeVar.t

module ScopeVarMap : Stdlib.Map.S with type key = ScopeVar.t

module StructName = Dcalc.Ast.StructName

module StructMap = Dcalc.Ast.StructMap

module StructFieldName = Dcalc.Ast.StructFieldName

module StructFieldMap : Stdlib.Map.S with type key = StructFieldName.t

module StructFieldMapLift : sig ... end

module EnumName = Dcalc.Ast.EnumName

module EnumMap = Dcalc.Ast.EnumMap

module EnumConstructor = Dcalc.Ast.EnumConstructor

module EnumConstructorMap : Stdlib.Map.S with type key = EnumConstructor.t

module EnumConstructorMapLift : sig ... end

type location =

| ScopeVar of ScopeVar.t Utils.Pos.marked

| SubScopeVar of ScopeName.t * SubScopeName.t Utils.Pos.marked * ScopeVar.t Utils.Pos.marked

module LocationSet : Stdlib.Set.S with type elt = location Utils.Pos.marked

Abstract syntax tree

type typ =

| TLit of Dcalc.Ast.typ_lit

| TStruct of StructName.t

| TEnum of EnumName.t

| TArrow of typ Utils.Pos.marked * typ Utils.Pos.marked

| TArray of typ

| TAny

type expr =

| ELocation of location

| EVar of expr Bindlib.var Utils.Pos.marked

| EStruct of StructName.t * expr Utils.Pos.marked StructFieldMap.t

| EStructAccess of expr Utils.Pos.marked * StructFieldName.t * StructName.t

| EEnumInj of expr Utils.Pos.marked * EnumConstructor.t * EnumName.t

| EMatch of expr Utils.Pos.marked * EnumName.t * expr Utils.Pos.marked EnumConstructorMap.t

| ELit of Dcalc.Ast.lit

| EAbs of ( expr, expr Utils.Pos.marked ) Bindlib.mbinder Utils.Pos.marked * typ Utils.Pos.marked list

| EApp of expr Utils.Pos.marked * expr Utils.Pos.marked list

| EOp of Dcalc.Ast.operator

| EDefault of expr Utils.Pos.marked list * expr Utils.Pos.marked * expr Utils.Pos.marked

| EIfThenElse of expr Utils.Pos.marked * expr Utils.Pos.marked * expr Utils.Pos.marked

| EArray of expr Utils.Pos.marked list

| ErrorOnEmpty of expr Utils.Pos.marked

The expressions use the Bindlib library, based on higher-order abstract syntax

val locations_used : expr Utils.Pos.marked -> LocationSet.t

type io_input =

| NoInput	(*
For an internal variable defined only in the scope, and does not appear in the input.
*)

| OnlyInput	(*
For variables that should not be redefined in the scope, because they appear in the input.
*)

| Reentrant	(*
For variables defined in the scope that can also be redefined by the caller as they appear in the input.
*)

This type characterizes the three levels of visibility for a given scope variable with regards to the scope's input and possible redefinitions inside the scope..

type io = {

io_output : bool Utils.Pos.marked;	(*
true is present in the output of the scope.
*)

io_input : io_input Utils.Pos.marked;

}

Characterization of the input/output status of a scope variable.

type rule =

| Definition of location Utils.Pos.marked * typ Utils.Pos.marked * io * expr Utils.Pos.marked

| Assertion of expr Utils.Pos.marked

| Call of ScopeName.t * SubScopeName.t

type scope_decl = {

scope_decl_name : ScopeName.t;

scope_sig : (typ Utils.Pos.marked * io) ScopeVarMap.t;

scope_decl_rules : rule list;

}

type struct_ctx = (StructFieldName.t * typ Utils.Pos.marked) list StructMap.t

type enum_ctx = (EnumConstructor.t * typ Utils.Pos.marked) list EnumMap.t

type program = {

program_scopes : scope_decl ScopeMap.t;

program_enums : enum_ctx;

program_structs : struct_ctx;

}

Variable helpers

module Var : sig ... end

module VarMap : Stdlib.Map.S with type key = Var.t

type vars = expr Bindlib.mvar

val make_var : Var.t Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box

val make_abs : vars -> expr Utils.Pos.marked Bindlib.box -> Utils.Pos.t -> typ Utils.Pos.marked list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_app : expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box list -> Utils.Pos.t -> expr Utils.Pos.marked Bindlib.box

val make_let_in : Var.t -> typ Utils.Pos.marked -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box -> expr Utils.Pos.marked Bindlib.box