Module Scopelang.Dependency

Scope dependencies

module SDependencies : Graph.Sig.P with type SDependencies.V.t = Ast.ScopeName.t and type SDependencies.E.label = Utils.Pos.t

On the edges, the label is the expression responsible for the use of the function

val build_program_dep_graph : Ast.program -> SDependencies.t

val check_for_cycle_in_scope : SDependencies.t -> unit

val get_scope_ordering : SDependencies.t -> Ast.ScopeName.t list

Type dependencies

module TVertex : sig ... end

module TVertexSet : Stdlib.Set.S with type TVertexSet.elt = TVertex.t

module TDependencies : Graph.Sig.P with type TDependencies.V.t = TVertex.t and type TDependencies.E.label = Utils.Pos.t

On the edges, the label is the expression responsible for the use of the function

val get_structs_or_enums_in_type : Ast.typ Utils.Pos.marked -> TVertexSet.t

val build_type_graph : Ast.struct_ctx -> Ast.enum_ctx -> TDependencies.t

val check_type_cycles : Ast.struct_ctx -> Ast.enum_ctx -> TVertex.t list