Module Shared_ast.Expr

Functions handling the expressions of shared_ast

Boxed constructors

val box : ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val evar : ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Var.t -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val etuple : ( ( [< Shared_ast__Definitions.dcalc | Shared_ast__Definitions.lcalc ] as 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box list -> Shared_ast__Definitions.StructName.t option -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val etupleaccess : ( ( [< Shared_ast__Definitions.dcalc | Shared_ast__Definitions.lcalc ] as 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> int -> Shared_ast__Definitions.StructName.t option -> Shared_ast__Definitions.naked_typ Utils.Marked.pos list -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val einj : ( ( [< Shared_ast__Definitions.dcalc | Shared_ast__Definitions.lcalc ] as 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> int -> Shared_ast__Definitions.EnumName.t -> Shared_ast__Definitions.naked_typ Utils.Marked.pos list -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val ematch : ( ( [< Shared_ast__Definitions.dcalc | Shared_ast__Definitions.lcalc ] as 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box list -> Shared_ast__Definitions.EnumName.t -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val earray : ( ( 'a Shared_ast__Definitions.any, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box list -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val elit : 'a Shared_ast__Definitions.any Shared_ast__Definitions.glit -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val eabs : ( ( 'a Shared_ast__Definitions.any, 't ) Shared_ast__Definitions.naked_gexpr, ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t ) Bindlib.mbinder Bindlib.box -> Shared_ast__Definitions.naked_typ Utils.Marked.pos list -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val eapp : ( ( 'a Shared_ast__Definitions.any, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box list -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val eassert : ( ( [< Shared_ast__Definitions.dcalc | Shared_ast__Definitions.lcalc ] as 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val eop : Shared_ast__Definitions.operator -> 't -> ( ( _ Shared_ast__Definitions.any, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val edefault : ( ( [< Shared_ast__Definitions.desugared | Shared_ast__Definitions.scopelang | Shared_ast__Definitions.dcalc ] as 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box list -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val eifthenelse : ( ( 'a Shared_ast__Definitions.any, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val eerroronempty : ( ( [< Shared_ast__Definitions.desugared | Shared_ast__Definitions.scopelang | Shared_ast__Definitions.dcalc ] as 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val ecatch : ( ( [ `Lcalc ], 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> Shared_ast__Definitions.except -> ( ( [ `Lcalc ], 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> 't -> ( ( [ `Lcalc ], 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val eraise : Shared_ast__Definitions.except -> 't -> ( ( [ `Lcalc ], 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

Manipulation of marks

val no_mark : 'm Shared_ast__Definitions.mark -> 'm Shared_ast__Definitions.mark

val mark_pos : 'm Shared_ast__Definitions.mark -> Utils.Pos.t

val pos : ( ( 'e, _ Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, _ Shared_ast__Definitions.mark ) Utils.Marked.t -> Utils.Pos.t

val ty : ( _, Shared_ast__Definitions.typed Shared_ast__Definitions.mark ) Utils.Marked.t -> Shared_ast__Definitions.naked_typ Utils.Marked.pos

val with_ty : Shared_ast__Definitions.naked_typ Utils.Marked.pos -> ( 'a, _ Shared_ast__Definitions.mark ) Utils.Marked.t -> ( 'a, Shared_ast__Definitions.typed Shared_ast__Definitions.mark ) Utils.Marked.t

val map_mark : ( Utils.Pos.t -> Utils.Pos.t ) -> ( Shared_ast__Definitions.naked_typ Utils.Marked.pos -> Shared_ast__Definitions.naked_typ Utils.Marked.pos ) -> 'm Shared_ast__Definitions.mark -> 'm Shared_ast__Definitions.mark

val map_mark2 : ( Utils.Pos.t -> Utils.Pos.t -> Utils.Pos.t ) -> ( Shared_ast__Definitions.typed -> Shared_ast__Definitions.typed -> Shared_ast__Definitions.naked_typ Utils.Marked.pos ) -> 'm Shared_ast__Definitions.mark -> 'm Shared_ast__Definitions.mark -> 'm Shared_ast__Definitions.mark

val fold_marks : ( Utils.Pos.t list -> Utils.Pos.t ) -> ( Shared_ast__Definitions.typed list -> Shared_ast__Definitions.naked_typ Utils.Marked.pos ) -> 'm Shared_ast__Definitions.mark list -> 'm Shared_ast__Definitions.mark

val untype : ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t -> ( ( 'a, Shared_ast__Definitions.untyped Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, Shared_ast__Definitions.untyped Shared_ast__Definitions.mark ) Utils.Marked.t Bindlib.box

Traversal functions

val map : 'ctx -> f: ( 'ctx -> ( ( 'a, 't1 ) Shared_ast__Definitions.naked_gexpr, 't1 ) Utils.Marked.t -> ( ( 'a, 't2 ) Shared_ast__Definitions.naked_gexpr, 't2 ) Utils.Marked.t Bindlib.box ) -> ( ( 'a, 't1 ) Shared_ast__Definitions.naked_gexpr, 't2 ) Utils.Marked.t -> ( ( 'a, 't2 ) Shared_ast__Definitions.naked_gexpr, 't2 ) Utils.Marked.t Bindlib.box

Flat (non-recursive) mapping on expressions.

If you want to apply a map transform to an expression, you can save up writing a painful match over all the cases of the AST. For instance, if you want to remove all errors on empty, you can write

let remove_error_empty =
  let rec f () e =
    match Marked.unmark e with
    | ErrorOnEmpty e1 -> Expr.map () f e1
    | _ -> Expr.map () f e
  in
  f () e

The first argument of map_expr is an optional context that you can carry around during your map traversal.

val map_top_down : f: ( ( ( 'a, 't1 ) Shared_ast__Definitions.naked_gexpr, 't1 ) Utils.Marked.t -> ( ( 'a, 't1 ) Shared_ast__Definitions.naked_gexpr, 't2 ) Utils.Marked.t ) -> ( ( 'a, 't1 ) Shared_ast__Definitions.naked_gexpr, 't1 ) Utils.Marked.t -> ( ( 'a, 't2 ) Shared_ast__Definitions.naked_gexpr, 't2 ) Utils.Marked.t Bindlib.box

Recursively applies f to the nodes of the expression tree. The type returned by f is hybrid since the mark at top-level has been rewritten, but not yet the marks in the subtrees.

val map_marks : f:( 't1 -> 't2 ) -> ( ( 'a, 't1 ) Shared_ast__Definitions.naked_gexpr, 't1 ) Utils.Marked.t -> ( ( 'a, 't2 ) Shared_ast__Definitions.naked_gexpr, 't2 ) Utils.Marked.t Bindlib.box

Expression building helpers

val make_var : (( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Var.t * 'b) -> (( 'a, 't ) Shared_ast__Definitions.naked_gexpr * 'b) Bindlib.box

val make_abs : ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Var.vars -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box -> Shared_ast__Definitions.naked_typ Utils.Marked.pos list -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

val make_app : ( ( 'a Shared_ast__Definitions.any, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Bindlib.box -> ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Bindlib.box list -> 'm Shared_ast__Definitions.mark -> ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Bindlib.box

val empty_thunked_term : 'm Shared_ast__Definitions.mark -> ( ( [< Shared_ast__Definitions.dcalc | Shared_ast__Definitions.desugared | Shared_ast__Definitions.scopelang ], 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t

val make_let_in : ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Var.t -> Shared_ast__Definitions.naked_typ Utils.Marked.pos -> ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Bindlib.box -> ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Bindlib.box -> Utils.Pos.t -> ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Bindlib.box

val make_multiple_let_in : ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Var.vars -> Shared_ast__Definitions.naked_typ Utils.Marked.pos list -> ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Bindlib.box list -> ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Bindlib.box -> Utils.Pos.t -> ( ( 'a, 'm Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, 'm Shared_ast__Definitions.mark ) Utils.Marked.t Bindlib.box

val make_default : ( ( [< Shared_ast__Definitions.desugared | Shared_ast__Definitions.scopelang | Shared_ast__Definitions.dcalc ] as 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t list -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> 't -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t

make_default ?pos exceptions just cons builds a term semantically equivalent to <exceptions | just :- cons> (the EDefault constructor) while avoiding redundant nested constructions. The position is extracted from just by default.

Note that, due to the simplifications taking place, the result might not be of the form EDefault:

<true :- x> is rewritten as x

<ex | true :- def>, when def is a default term <j :- c> without exceptions, is collapsed into <ex | def>

<ex | false :- _>, when ex is a single exception, is rewritten as ex

Transformations

val remove_logging_calls : ( ( 'a Shared_ast__Definitions.any, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Bindlib.box

Removes all calls to Log unary operators in the AST, replacing them by their argument.

val format : ?debug:bool -> Shared_ast__Definitions.decl_ctx -> Stdlib.Format.formatter -> ( ( _, _ Shared_ast__Definitions.mark ) Shared_ast__Definitions.naked_gexpr, _ Shared_ast__Definitions.mark ) Utils.Marked.t -> unit

Analysis and tests

val equal_lit : 'a Shared_ast__Definitions.glit -> 'a Shared_ast__Definitions.glit -> bool

val compare_lit : 'a Shared_ast__Definitions.glit -> 'a Shared_ast__Definitions.glit -> int

val equal_location : 'a Shared_ast__Definitions.glocation Utils.Marked.pos -> 'a Shared_ast__Definitions.glocation Utils.Marked.pos -> bool

val compare_location : 'a Shared_ast__Definitions.glocation Utils.Marked.pos -> 'a Shared_ast__Definitions.glocation Utils.Marked.pos -> int

val equal : ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> bool

Determines if two expressions are equal, omitting their position information

val compare : ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> int

Standard comparison function, suitable for e.g. Set.Make. Ignores position information

val compare_typ : Shared_ast__Definitions.naked_typ Utils.Marked.pos -> Shared_ast__Definitions.naked_typ Utils.Marked.pos -> int

val is_value : ( ( 'a Shared_ast__Definitions.any, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> bool

val free_vars : ( ( 'a Shared_ast__Definitions.any, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t Var.Set.t

val size : ( ( 'a, 't ) Shared_ast__Definitions.naked_gexpr, 't ) Utils.Marked.t -> int

Used by the optimizer to know when to stop