Module Shared_ast.Op

Classification of operators on how they should be typed

type monomorphic =

| Monomorphic	(*
Operands and return types of the operator are fixed
*)

type polymorphic =

| Polymorphic	(*
The operator is truly polymorphic: it's the same runtime function that may work on multiple types. We require that resolving the argument types from right to left trivially resolves all type variables declared in the operator type.
*)

type overloaded =

| Overloaded	(*
The operator is ambiguous and requires the types of its arguments to be known before it can be typed, using a pre-defined table
*)

type resolved =

| Resolved	(*
Explicit monomorphic versions of the overloaded operators
*)

type (_, _) kind =

| Monomorphic : ( 'a any, monomorphic ) kind

| Polymorphic : ( 'a any, polymorphic ) kind

| Overloaded : ( [< desugared ], overloaded ) kind

| Resolved : ( [< scopelang | dcalc | lcalc ], resolved ) kind

Classification of operators. This could be inlined in the definition of t but is more concise this way

type (_, _) t =

| Not : ( 'a any, monomorphic ) t

| GetDay : ( 'a any, monomorphic ) t

| GetMonth : ( 'a any, monomorphic ) t

| GetYear : ( 'a any, monomorphic ) t

| FirstDayOfMonth : ( 'a any, monomorphic ) t

| LastDayOfMonth : ( 'a any, monomorphic ) t

| Length : ( 'a any, polymorphic ) t

| Log : log_entry * Catala_utils.Uid.MarkedString.info list -> ( 'a any, polymorphic ) t

| Minus : ( desugared, overloaded ) t

| Minus_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Minus_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Minus_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Minus_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| ToRat : ( desugared, overloaded ) t

| ToRat_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| ToRat_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| ToMoney : ( desugared, overloaded ) t

| ToMoney_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Round : ( desugared, overloaded ) t

| Round_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Round_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| And : ( 'a any, monomorphic ) t

| Or : ( 'a any, monomorphic ) t

| Xor : ( 'a any, monomorphic ) t

| Eq : ( 'a any, polymorphic ) t

| Map : ( 'a any, polymorphic ) t

| Concat : ( 'a any, polymorphic ) t

| Filter : ( 'a any, polymorphic ) t

| Reduce : ( 'a any, polymorphic ) t

| Add : ( desugared, overloaded ) t

| Add_int_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Add_rat_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Add_mon_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Add_dat_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Add_dur_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Sub : ( desugared, overloaded ) t

| Sub_int_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Sub_rat_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Sub_mon_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Sub_dat_dat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Sub_dat_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Sub_dur_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Mult : ( desugared, overloaded ) t

| Mult_int_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Mult_rat_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Mult_mon_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Mult_dur_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Div : ( desugared, overloaded ) t

| Div_int_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Div_rat_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Div_mon_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Div_mon_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Div_dur_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Lt : ( desugared, overloaded ) t

| Lt_int_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Lt_rat_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Lt_mon_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Lt_dat_dat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Lt_dur_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Lte : ( desugared, overloaded ) t

| Lte_int_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Lte_rat_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Lte_mon_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Lte_dat_dat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Lte_dur_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Gt : ( desugared, overloaded ) t

| Gt_int_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Gt_rat_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Gt_mon_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Gt_dat_dat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Gt_dur_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Gte : ( desugared, overloaded ) t

| Gte_int_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Gte_rat_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Gte_mon_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Gte_dat_dat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Gte_dur_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Eq_int_int : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Eq_rat_rat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Eq_mon_mon : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Eq_dur_dur : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Eq_dat_dat : ( [< scopelang | dcalc | lcalc ], resolved ) t

| Fold : ( 'a any, polymorphic ) t