Module Print.UserFacing

User-facing, localised printer

val unit : Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> Runtime_ocaml.Runtime.unit -> unit

val bool : Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> Runtime_ocaml.Runtime.bool -> unit

val integer : Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> Runtime_ocaml.Runtime.integer -> unit

val decimal : Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> Runtime_ocaml.Runtime.decimal -> unit

val money : Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> Runtime_ocaml.Runtime.money -> unit

val date : Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> Runtime_ocaml.Runtime.date -> unit

val duration : Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> Runtime_ocaml.Runtime.duration -> unit

val lit : Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> Shared_ast__Definitions.lit -> unit

val lit_to_string : Catala_utils.Cli.backend_lang -> Shared_ast__Definitions.lit -> string

val value : ?fallback: ( Stdlib.Format.formatter -> ( ( 'a, 'a, 't ) Shared_ast__Definitions.base_gexpr, 't Shared_ast__Definitions.mark ) Catala_utils.Mark.ed -> unit ) -> Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> ( ( 'a, 'a, 't ) Shared_ast__Definitions.base_gexpr, 't Shared_ast__Definitions.mark ) Catala_utils.Mark.ed -> unit

Prints a value in a localised format, intended to be read by an end-user.

parameter fallback
is called upon non-value expressions (by default, Invalid_argument is raised)

val expr : Catala_utils.Cli.backend_lang -> Stdlib.Format.formatter -> ( ( _, _, _ ) Shared_ast__Definitions.base_gexpr, _ Shared_ast__Definitions.mark ) Catala_utils.Mark.ed -> unit

This combines the user-facing value printer and the generic expression printer to handle all AST nodes