Module Surface.Name_resolution

Builds a context that allows for mapping each name to a precise uid, taking lexical scopes into account

Name resolution context

type ident = string

type unique_rulename =

| Ambiguous of Utils.Pos.t list

| Unique of Desugared.Ast.RuleName.t Utils.Marked.pos

type scope_def_context = {

default_exception_rulename : unique_rulename option;

label_idmap : Desugared.Ast.LabelName.t Desugared.Ast.IdentMap.t;

}

type scope_context = {

var_idmap : Shared_ast.ScopeVar.t Desugared.Ast.IdentMap.t;	(*
Scope variables
*)

scope_defs_contexts : scope_def_context Desugared.Ast.ScopeDefMap.t;	(*
What is the default rule to refer to for unnamed exceptions, if any
*)

sub_scopes_idmap : Shared_ast.SubScopeName.t Desugared.Ast.IdentMap.t;	(*
Sub-scopes variables
*)

sub_scopes : Shared_ast.ScopeName.t Scopelang.Ast.SubScopeMap.t;	(*
To what scope sub-scopes refer to?
*)

}

Inside a scope, we distinguish between the variables and the subscopes.

type struct_context = Shared_ast.typ Shared_ast.StructFieldMap.t

Types of the fields of a struct

type enum_context = Shared_ast.typ Shared_ast.EnumConstructorMap.t

Types of the payloads of the cases of an enum

type var_sig = {

var_sig_typ : Shared_ast.typ;

var_sig_is_condition : bool;

var_sig_io : Ast.scope_decl_context_io;

var_sig_states_idmap : Shared_ast.StateName.t Desugared.Ast.IdentMap.t;

var_sig_states_list : Shared_ast.StateName.t list;

}

type context = {

local_var_idmap : Desugared.Ast.expr Shared_ast.Var.t Desugared.Ast.IdentMap.t;	(*
Inside a definition, local variables can be introduced by functions arguments or pattern matching
*)

scope_idmap : Shared_ast.ScopeName.t Desugared.Ast.IdentMap.t;	(*
The names of the scopes
*)

struct_idmap : Shared_ast.StructName.t Desugared.Ast.IdentMap.t;	(*
The names of the structs
*)

field_idmap : Shared_ast.StructFieldName.t Shared_ast.StructMap.t Desugared.Ast.IdentMap.t;	(*
The names of the struct fields. Names of fields can be shared between different structs
*)

enum_idmap : Shared_ast.EnumName.t Desugared.Ast.IdentMap.t;	(*
The names of the enums
*)

constructor_idmap : Shared_ast.EnumConstructor.t Shared_ast.EnumMap.t Desugared.Ast.IdentMap.t;	(*
The names of the enum constructors. Constructor names can be shared between different enums
*)

scopes : scope_context Scopelang.Ast.ScopeMap.t;	(*
For each scope, its context
*)

structs : struct_context Shared_ast.StructMap.t;	(*
For each struct, its context
*)

enums : enum_context Shared_ast.EnumMap.t;	(*
For each enum, its context
*)

var_typs : var_sig Shared_ast.ScopeVarMap.t;	(*
The signatures of each scope variable declared
*)

}

Main context used throughout Surface.Desugaring

Helpers

val raise_unsupported_feature : string -> Utils.Pos.t -> 'a

Temporary function raising an error message saying that a feature is not supported yet

val raise_unknown_identifier : string -> ident Utils.Marked.pos -> 'a

Function to call whenever an identifier used somewhere has not been declared in the program previously

val get_var_typ : context -> Shared_ast.ScopeVar.t -> Shared_ast.typ

Gets the type associated to an uid

val is_var_cond : context -> Shared_ast.ScopeVar.t -> bool

val get_var_io : context -> Shared_ast.ScopeVar.t -> Ast.scope_decl_context_io

val get_var_uid : Shared_ast.ScopeName.t -> context -> ident Utils.Marked.pos -> Shared_ast.ScopeVar.t

Get the variable uid inside the scope given in argument

val get_subscope_uid : Shared_ast.ScopeName.t -> context -> ident Utils.Marked.pos -> Shared_ast.SubScopeName.t

Get the subscope uid inside the scope given in argument

val is_subscope_uid : Shared_ast.ScopeName.t -> context -> ident -> bool

is_subscope_uid scope_uid ctxt y returns true if y belongs to the subscopes of scope_uid.

val belongs_to : context -> Shared_ast.ScopeVar.t -> Shared_ast.ScopeName.t -> bool

Checks if the var_uid belongs to the scope scope_uid

val get_def_typ : context -> Desugared.Ast.ScopeDef.t -> Shared_ast.typ

Retrieves the type of a scope definition from the context

val is_def_cond : context -> Desugared.Ast.ScopeDef.t -> bool

val is_type_cond : Ast.typ -> bool

val add_def_local_var : context -> ident -> context * Desugared.Ast.expr Shared_ast.Var.t

Adds a binding to the context

val get_def_key : Ast.qident -> Ast.ident Utils.Marked.pos option -> Shared_ast.ScopeName.t -> context -> Utils.Pos.t -> Desugared.Ast.ScopeDef.t

Usage: get_def_key var_name var_state scope_uid ctxt pos

API

val form_context : Ast.program -> context

Derive the context from metadata, in one pass over the declarations