Module Utils.Cli

type frontend_lang = [

| `En

| `Fr

| `NonVerbose

]

type backend_lang = [

| `En

| `Fr

]

val to_backend_lang : frontend_lang -> backend_lang

Configuration globals

val source_files : string list Stdlib.ref

Source files to be compiled

val locale_lang : backend_lang Stdlib.ref

val contents : string Stdlib.ref

val debug_flag : bool Stdlib.ref

val style_flag : bool Stdlib.ref

Styles the terminal output

val optimize_flag : bool Stdlib.ref

val max_prec_digits : int Stdlib.ref

Max number of digits to show for decimal results

val trace_flag : bool Stdlib.ref

CLI terms

val file : string Cmdliner.Term.t

val debug : bool Cmdliner.Term.t

val unstyled : bool Cmdliner.Term.t

val trace_opt : bool Cmdliner.Term.t

val wrap_weaved_output : bool Cmdliner.Term.t

val backend : string Cmdliner.Term.t

type backend_option =

| Latex

| Makefile

| Html

| Run

| OCaml

val language : string option Cmdliner.Term.t

val max_prec_digits_opt : int option Cmdliner.Term.t

val ex_scope : string option Cmdliner.Term.t

val output : string option Cmdliner.Term.t

val catala_t : (string -> bool -> bool -> bool -> bool -> string -> string option -> int option -> bool -> bool -> string option -> string option -> 'a) -> 'a Cmdliner.Term.t

Main entry point: catala_t file debug unstyled wrap_weaved_output backend language max_prec_digits_opt trace_opt optimize ex_scope output

val version : string

val info : Cmdliner.Term.info

Terminal formatting

Markers

val print_with_style : ANSITerminal.style list -> ('a, unit, string) Stdlib.format -> 'a

val debug_marker : unit -> string

val error_marker : unit -> string

val warning_marker : unit -> string

val result_marker : unit -> string

val log_marker : unit -> string

Printers

val concat_with_line_depending_prefix_and_suffix : (int -> string) -> (int -> string) -> string list -> string

val add_prefix_to_each_line : string -> (int -> string) -> string

The int argument of the prefix corresponds to the line number, starting at 0

val debug_print : string -> unit

val error_print : string -> unit

val warning_print : string -> unit

val result_print : string -> unit

val log_print : string -> unit