Module Uid.MarkedString

The only kind of information carried in Catala identifiers is the original string of the identifier annotated with the position where it is declared or used.

type info = string Pos.marked

val format_info : Stdlib.Format.formatter -> info -> unit